% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_q.R
\name{add_CPUE}
\alias{add_CPUE}
\title{Add a q setup line into an SS control file}
\usage{
add_CPUE(ctl.in, ctl.out = NULL, overwrite = FALSE,
  q = data.frame(fleet = 3, link = 1, link_info = 0, extra_se = 0,
  biasadj = 0, float = 0, LO = -20, HI = 20, INIT = 0, PRIOR = 0, PR_SD =
  99, PR_type = 0, PHASE = 1, env_var = 0, use_dev = 0, dev_mnyr = 0,
  dev_mxyr = 0, dev_PH = 0, Block = 0, Blk_Fxn = 0, name = NULL))
}
\arguments{
\item{ctl.in}{An SS control file name to read in.}

\item{ctl.out}{The SS control file to read out.}

\item{overwrite}{Logical. Overwrite an existing file with the same name as
\code{ctl.out}?}

\item{q}{a dataframe containing the q parameter lines to add.}
}
\value{
A modified SS control file.
}
\description{
This function adds a q setup line to an SS 3.30 control file
}
\author{
Kelli Johnson
}
