% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/add_nulls.R
\name{add_nulls}
\alias{add_nulls}
\title{Add NULL values to non-existent list elements}
\usage{
add_nulls(param_list, desired_params)
}
\arguments{
\item{param_list}{A list in which the names correspond to parameter names
and the values correspond to the values to be passed.}

\item{desired_params}{A character vector of desired list elements.}
}
\value{
A list with the desired elements as described by the
  \code{desired_params} argument. Any values that were missing in
  \code{param_list} will be returned with values of \code{NULL}.
}
\description{
Add NULL values to non-existent list elements
}
\author{
Sean C. Anderson
}

