% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_tv.r
\name{add_tv_parlines}
\alias{add_tv_parlines}
\title{Add short time varying parameter lines. At time of writing, this method will
work for MG, selectivity, and catchability time varying, but not for SR}
\usage{
add_tv_parlines(string, tab, ctl_string, ss3.ctl)
}
\arguments{
\item{string}{The code representing the section the parameter is from.}

\item{tab}{As created in \code{change_tv()}}

\item{ctl_string}{The code as called in the .ss_new comment for time varying.}

\item{ss3.ctl}{A ss control file that has been read in using \code{readLines()}.}
}
\value{
A modified version of ss3.ctl (a vector of strings), containing the
new parameter line
}
\description{
Add short time varying parameter lines. At time of writing, this method will
work for MG, selectivity, and catchability time varying, but not for SR
}
