% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bias.R
\name{calculate_bias}
\alias{calculate_bias}
\title{Calculate bias adjustment for recruitment deviations}
\usage{
calculate_bias(dir, ctl_file_in)
}
\arguments{
\item{dir}{A character string specifying the path to the folder with the
results from the preliminary run of the estimation model.}

\item{ctl_file_in}{A string providing the path to the input SS \code{.ctl} file.}
}
\value{
A list of bias adjustment parameters.
}
\description{
Bias adjustment is performed to ensure that the only the most informative
data available are used when estimating recruitment deviations.
This process involves running the estimation method with the hessian
both this function and then running the estimation method again with
the new values.
}
\details{
\itemize{
\item Estimate recruitment and the standard error about those estimates.
\item Correct the estimates given their estimated uncertainty using a ramp.
\item Save a new control file.
\item Estimate model parameters.
}
}
\author{
Kelli Faye Johnson
}
