% Generated by roxygen2 (4.0.0): do not edit by hand
\name{calculate_data_units}
\alias{calculate_data_units}
\title{Given sampling arguments, calculate super set of fleets, years, and data
types.}
\usage{
calculate_data_units(index_params = NULL, lcomp_params = NULL,
  agecomp_params = NULL, calcomp_params = NULL, mlacomp_params = NULL)
}
\arguments{
  \item{index_params,}{lcomp_params, agecomp_params,
  calcomp_params, mlacomp_params Named lists containing the
  arguments for the different sampling functions. #'}
}
\description{
Given sampling arguments, calculate super set of fleets,
years, and data types.
}
\examples{
## Should throw error since nothing passed
calculate_data_units()
## Only one fleet
calculate_data_units(lcomp_params=list(fleets=1, years=c(3,4,6)))
## Add new fleet
calculate_data_units(lcomp_params=list(fleets=1, years=c(3,4,6)),
                     agecomp_params=list(fleets=2, years=5))
## If CAL data called, need other types even if not specified
calculate_data_units(calcomp_params=list(fleets=1, years=c(3,4,6)))
}
\author{
Cole Monnahan
}

