% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-write.R
\name{case_binning}
\alias{case_binning}
\title{Write a case file for binning to the disk.}
\usage{
case_binning(lbin_method = NULL, pop_binwidth = NULL,
  pop_minimum_size = NULL, pop_maximum_size = NULL, bin_vector = NULL,
  case, spp)
}
\arguments{
\item{lbin_method}{xxxxxxx}

\item{pop_binwidth}{Population length bin width. Note that this value must 
be smaller than the bin width specified in length composition data len_bins or 
SS3 will fail (see notes in the SS3 manual).}

\item{pop_minimum_size}{Population minimum length bin value.}

\item{pop_maximum_size}{Population maximum length bin value.}

\item{bin_vector}{xxxx}

\item{case}{The case number you want to write to.
If \code{case = 1}, then the result will be \code{'index1'}.}

\item{spp}{A vector of character values argument specifying the species.}
}
\description{
Use \R code to write arguments to the disk, which
will later be used in a \pkg{ss3sim} simulation.
}
