% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-write.R
\name{case_comp_caal}
\alias{case_comp_caal}
\title{Write a case file for conditional length-at-age composition data}
\usage{
case_comp_caal(fleets = 1, Nsamp = NULL, years = NULL,
  random_gen = NULL, fixed.number = NULL, method = NULL, Nhauls = NULL,
  ESS = NULL, change_ess_now = NULL, fit.on.agecomp = NULL, rw = NULL,
  percent_read = 100, max_ss = NULL, type, case, spp)
}
\arguments{
\item{fleets}{Vector of fleet numbers, where the order of
\code{fleets} will dictate the order of all remaining arguments.}

\item{Nsamp}{A list of length \code{length(fleets)},
where each element of the list contains a vector of
sample sizes for each year for that given fleet.}

\item{years}{A list of length \code{length(fleets)},
where each element of the list contains a vector of
years for the given fleet.}

\item{random_gen}{if randomly generated numbers}

\item{fixed.number}{If the sampling should be length-stratified with
a fixed number of ages at length}

\item{method}{If it should be simple random sampling or length-stratified}

\item{Nhauls}{Number of hauls to sample age data from}

\item{ESS}{Change effective samplign size to number of hauls}

\item{change_ess_now}{now or later (at the end of sampling in ss3sim_base)}

\item{fit.on.agecomp}{Fit on marginal age composition instead of default
conditional age at length}

\item{rw}{Run iterative re-weighing (McAllister and Ianelli)}

\item{percent_read}{collected vs read}

\item{max_ss}{max number read (use either this or percent_read)}

\item{type}{A character value of \code{"agecomp"} or \code{"lcomp"},
to write age- or length-composition specifications, respectively.
Argument can be a vector (e.g., \code{c("agecomp", "lcomp")}) if you want
the case files to be the same for length and age compositions.}

\item{case}{The casenumber you want to write to.
If \code{case = 1} and \code{type = "agecomp"},
then the result will be \code{'agecomp1'}.}

\item{spp}{A vector of character values argument specifying the species.}
}
\description{
Use \R code to write arguments to the disk, which
will later be used in a \pkg{ss3sim} simulation.
}
