% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-write.R
\name{case_data}
\alias{case_data}
\title{Write a case file for data to the disk.}
\usage{
case_data(age_bins = NULL, len_bins = NULL, pop_binwidth = NULL,
  pop_minimum_size = NULL, pop_maximum_size = NULL, lcomp_constant = NULL,
  tail_compression = NULL, case, spp)
}
\arguments{
\item{age_bins}{A numeric vector of age bins to use. If left as NULL 
then the age bin structure will be taken from the OM}

\item{len_bins}{A numeric vector of length bins to use. If left as NULL 
then the length bin structure will be taken from the OM}

\item{pop_binwidth}{Population length bin width. Note that this value must 
be smaller than the bin width specified in length composition data len_bins or 
SS3 will fail (see notes in the SS3 manual).}

\item{pop_minimum_size}{Population minimum length bin value.}

\item{pop_maximum_size}{Population maximum length bin value.}

\item{lcomp_constant}{A new robustification constant for length composition data 
to be used. Must be a numeric value, as a proportion. For example 0.1 means 10 percent. 
See the SS3 manual for further information. A NULL value indicates no action resulting 
in using the current value, and a value of 0 will throw an error since that leads to an 
error when zeroes exist in the data. Instead use a very small value like 1e-07.}

\item{tail_compression}{A new tail compression value to be used in SS3. Must be a numeric 
value, as a proportion. For example 0.1 means 10 percent. See the SS3 manual for further 
information. A NULL value indicates no action, a negative value indicates to SS3 to 
ignore it (not use that feature).}

\item{case}{The case number you want to write to.
If \code{case = 1}, then the result will be \code{'index1'}.}

\item{spp}{A vector of character values argument specifying the species.}
}
\description{
Use \R code to write arguments to the disk, which
will later be used in a \pkg{ss3sim} simulation.
}
