% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case-write.R
\name{case_deparse}
\alias{case_deparse}
\title{Turn an argument describing an object into a character.}
\usage{
case_deparse(x)
}
\arguments{
\item{x}{The argument you would like to \code{deparse}.
\code{"M1-F1-D1-R1"}}
}
\value{
A single character value.
}
\description{
Turn an argument describing an object into a character.
}
\details{
Includes checks to make sure multiple lines will not be created.
}

