% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-write.R
\name{case_mla}
\alias{case_mla}
\title{Write a case file for mean length at age data to the disk.}
\usage{
case_mla(fleets = 1, years = NULL, Nsamp = NULL, ageplus = NULL, case,
  spp)
}
\arguments{
\item{fleets}{Vector of fleet numbers, where the order of
\code{fleets} will dictate the order of all remaining arguments.}

\item{years}{A list of length \code{length(fleets)},
where each element of the list contains a vector of
years for the given fleet.}

\item{Nsamp}{The number of fish to measure for length and age}

\item{ageplus}{The age-plus group, if different from the population max age, in case \code{sample_wtatage}
is used after}

\item{case}{The case number you want to write to.
If \code{case = 1}, then the result will be \code{'index1'}.}

\item{spp}{A vector of character values argument specifying the species.}
}
\description{
Use \R code to write arguments to the disk, which
will later be used in a \pkg{ss3sim} simulation.
}
