% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-write.R
\name{case_wtatage}
\alias{case_wtatage}
\title{Write a case file for weight at age data to the disk.}
\usage{
case_wtatage(fleets = 1, years = NULL, cv_wtatage = 0.05,
  random_gen = NULL, case, spp)
}
\arguments{
\item{fleets}{Vector of fleet numbers, where the order of
\code{fleets} will dictate the order of all remaining arguments.}

\item{years}{A list of length \code{length(fleets)},
where each element of the list contains a vector of
years for the given fleet.}

\item{cv_wtatage}{A coefficient of variation}

\item{random_gen}{if numbers were randomly generated in case file}

\item{case}{The case number you want to write to.
If \code{case = 1}, then the result will be \code{'index1'}.}

\item{spp}{A vector of character values argument specifying the species.}
}
\description{
Use \R code to write arguments to the disk, which
will later be used in a \pkg{ss3sim} simulation.
}
