% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_a.r
\name{change_a}
\alias{change_a}
\title{Function to modify ageing error parameters in \code{.dat} files.}
\usage{
change_a(dat_list, outfile, df, agerr_mat, mis_spec = NULL,
  write_file = TRUE)
}
\arguments{
\item{dat_list}{Input SS3 data file}

\item{outfile}{Output SS3 data file}

\item{df}{Dataframe giving fleets, years, gender etc and starts with corresponding 
ageing error matrix index}

\item{agerr_mat}{new ageing error matrix to be added to existing ones and index 
number}

\item{mis_spec}{whether to mis-specify ageing error in the EM file 
(turns back to assuming no ageing error in the EM)}

\item{write_file}{A logical value indicating if the altered file should be
written to the disk. The default value is \code{TRUE}.}
}
\value{
Altered versions of SS3 \code{.dat} files are written to the disk.
}
\description{
Takes ss3 \code{.dat} files and changes the ageing error.
The function can be called by itself or within \code{\link{run_ss3sim}} to alter 
an operating and estimation model \code{.ctl} files. If used with \code{\link{run_ss3sim}} 
the case file should be named \code{A}. A suggested (default) case letter is \code{A} for 
Ageing error.
}
\details{
\code{change_a} modifies the ageing error in the \code{.dat} file both
in the OM and in the EM unless specified otherwise. There is no option yet to
specify different ageing errors in OM and EM, but can be easily implemented.
Using this case file in \code{\link{ss3sim_base}} means that the expected 
data out of the OM will include ageing error before the sampling occurs.
}
\section{Which arguments to specifiy in case files}{

All function argument descriptions that start with an asterisk (*) will be passed
through the case files to \code{\link{run_ss3sim}}. If one of these arguments
is not specified in a case file, then a value of \code{NULL} will be passed,
which may or may not be an appropriate value. Other arguments will be ignored
if specified.
}

\examples{
\dontrun{
# Find the SS3sim simple model in the package:
d         <- system.file("extdata", package = "ss3sim")
simple    <- paste0(d, "/Simple")
file.copy(simple, ".", recursive = TRUE)
setwd("simple")

wd <- getwd()

### ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~###
### READ EXISTING CASES OR CREATE CASES ###
### ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~###

## Read existing files

# Find the example data folders:
case_folder <- system.file("extdata", "new-cases", package = "ss3sim")

arg_list <- get_caseargs(case_folder, scenario = "D0-F0-A2-cod",
                         case_files = list(A = "A", 
                         D = c("index", "lcomp", "agecomp"), 
                         F ="F"))

### ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~###
### THIS IS THE FUNCTION ALONE - CAN BE RUN FOR ANY SCENARIO ABOVE ###
### ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~###

# Here we plot the selectivity curves for each case

setwd(wd)

datfile.orig <- SS_readdat("simple.dat", verbose = FALSE)

a_params <- arg_list$A
temp <- with(a_params,
     change_a(dat_list             =  datfile.orig,
              outfile              = "agerror.dat", 
              df                   =  df,
              agerr_mat            = agerr_mat,
              write_file           = FALSE,
              # mis_spec           = mis_spec
     ))

}
}
\seealso{
Other change functions: \code{\link{change_data}},
  \code{\link{change_em_binning}}, \code{\link{change_e}},
  \code{\link{change_f}}, \code{\link{change_maturity}},
  \code{\link{change_p}}, \code{\link{change_retro}},
  \code{\link{change_tv}}, \code{\link{change_x}}
}
\author{
Gwladys Lambert
}
