% Generated by roxygen2 (4.0.1): do not edit by hand
\name{change_bin}
\alias{change_bin}
\title{Change age or length bins in an SS operating model}
\usage{
change_bin(file_in, file_out, bin_vector, type = c("length", "age"),
  pop_bin = NULL, write_file = TRUE)
}
\arguments{
  \item{file_in}{A character value giving the location of
  an SS \code{.dat} file to input.}

  \item{file_out}{A character value giving the location of
  an SS \code{.dat} file to output.}

  \item{bin_vector}{A numeric vector of bins to substitute
  into the \code{.dat} file.}

  \item{type}{One of \code{"length"} or \code{"age"}. This
  controls whether the function acts on the age composition
  or length composition data. Default is \code{"length"}.}

  \item{pop_bin}{Choose a real number to choose the
  population bin width. This option only works if
  \code{"lbin_method"} is set to \code{"2"}. Default is
  \code{"NULL"} which leaves the original value.}

  \item{write_file}{Should the \code{.dat} file be written?
  The new \code{.dat} file will always be returned
  invisibly by the function. Setting \code{write_file =
  FALSE} can be useful for testing. Note that you must
  supply a value to the argument \code{file_out}, but this
  argument can be set to any arbitrary value (such as
  \code{NULL}) if \code{write_file = FALSE}.}
}
\description{
\code{change_bin} alters the bin structure for the age or
length composition data in an SS operating model. By adding
dummy data at the appropriate bin sizes to the SS
\code{.dat} file, SS will record age or length composition
data in the appropriate bins when the operating model is
run.
}
\examples{
d <- system.file("extdata", package = "ss3sim")
f_in <- paste0(d, "/example-om/data.ss_new")
l <- change_bin(f_in, file_out = NULL, type = "length",
  bin_vector = seq(2, 8, 2), write_file = FALSE)
print(l$lbin_vector)
print(head(l$lencomp))

a <- change_bin(f_in, file_out = NULL, type = "age",
  bin_vector = seq(2, 8, 2), write_file = FALSE)
print(a$agebin_vector)
print(head(a$agecomp))
}
\author{
Ian Taylor, Sean Anderson
}
\seealso{
\code{\link{sample_lcomp}}, \code{\link{sample_agecomp}}
}

