% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_catch.R
\name{change_catch}
\alias{change_catch}
\title{Change catch in the datafile}
\usage{
change_catch(dat_list, f_params)
}
\arguments{
\item{dat_list}{A data file as read in using `r4ss::SS_readdat()`}

\item{f_params}{A list of parameters related to fishing mortality.
The start and end year of the model will be based on
the years with specified catch.}
}
\value{
a modified data file as a list in R.
}
\description{
Change catch in the datafile so at least all combinations of fleet, seas, yr,
needed for catch are available
}
\author{
Kathryn Doering
}
