% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_dat_bin.R
\name{change_dat_bin}
\alias{change_dat_bin}
\title{Change the bins for a composition object}
\usage{
change_dat_bin(object, bins)
}
\arguments{
\item{object}{A data frame from a list object read in by \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}}.
The leading columns, typically \code{year}, \code{month}, ... can be anything,
but there must be columns with names that start with a lower-case letter,
followed by integers.}

\item{bins}{A vector of characters or whatever you want the names of the
new bins to be. Typically, this will be output from \code{\link[=setup_bins]{setup_bins()}}.}
}
\value{
A modified data frame where columns holding old composition data are
removed in their entirety and new columns of ones are filled for each value in
\code{bins}.
}
\description{
Change the bins of a data frame object from a \code{dat_list},
filling in the columns with ones.
}
