% Generated by roxygen2 (4.0.2): do not edit by hand
\name{change_data}
\alias{change_data}
\title{Change the data that is available as output from an SS operating model.}
\usage{
change_data(file_in, file_out, fleets = NULL, years = NULL, types = NULL,
  age_bins = NULL, len_bins = NULL, write_file = TRUE)
}
\arguments{
  \item{file_in}{A character value giving the location of
  an SS \code{.dat} file to input.}

  \item{file_out}{A character value giving the location of
  an SS \code{.dat} file to output.}

  \item{fleets}{A numeric vector of fleets}

  \item{years}{A numeric vector of years}

  \item{types}{A vector that can take the one or all of the
  following entries: \code{"len"}, \code{"age"},
  \code{"cal"}, \code{"mla"}. \code{types} controls what
  data structures the function acts on, with \code{"len"}
  augmenting the length composition data, \code{"age"}
  augmenting the age composition, \code{"cal"} augmenting
  the conditional age at length, and \code{"mla"}
  augmenting the mean length at age data.}

  \item{age_bins}{A numeric vector of age bins to use. If
  left as \code{NULL} then the age bin structure will be
  taken from the OM.}

  \item{len_bins}{A numeric vector of length bins to use.
  If left as \code{NULL} then the length bin structure will
  be taken from the OM.}

  \item{write_file}{Should the \code{.dat} file be written?
  The new \code{.dat} file will always be returned
  invisibly by the function. Setting \code{write_file =
  FALSE} can be useful for testing. Note that you must
  supply a value to the argument \code{file_out}, but this
  argument can be set to any arbitrary value (such as
  \code{NULL}) if \code{write_file = FALSE}.}
}
\description{
\code{change_data} alters the bin structure for the age or
length composition data in an SS operating model. Original
data is removed and dummy data is added at the appropriate
bin sizes to the SS \code{.dat} file. This causes SS to
record age or length composition data in the appropriate
bins when the operating model is run. Additionally,
\code{change_data} will introduce dummy conditional
length-at-age or size- (weight or length) at-age data to
the \code{.dat} file. For each data type altered,
\code{change_data} will add data in a full factorial
manner, working with existing fleets for all years;
potentially adding many rows of data. Currently,
\code{.dat} files with multiple genders cannot be
manipulated with \code{change_data}. Use
\code{\link{sample_lcomp}}, \code{\link{sample_agecomp}},
and \code{\link{sample_calcomp}} to reduce the data. It is
not intended for an \pkg{ss3sim} user to use
\code{change_data} directly. Instead \code{change_data} is
called internally by \code{\link{ss3sim_base}} based on
arguments to the sampling functions.
}
\examples{
## These examples are in development still and untested
d <- system.file("extdata", package = "ss3sim")
fleets <- 1:2
years <- c(5,10,15)
types <- c("len", "age")
file_in <- paste0(d, "/models/cod-om/codOM.dat")
## Basic test with just length data, default bins
out <- change_data(file_in,"test2.dat", types="len", years=years,
                   fleets=fleets, write_file=FALSE)
print(out$lbin_vector)
print(tail(out$lencomp[, 1:8]))
## Change the length bins
out <- change_data(file_in,"test.dat", types="len", years=years,
                   fleets=fleets, len_bins=3:6, write_file=FALSE)
print(out$lbin_vector)
print(tail(out$lencomp[, 1:8]))
## Change data types
out <- change_data(file_in,"test.dat", types=c("len", "age"), years=years,
                   fleets=fleets, len_bins=3:6, age_bins=5:7, write_file=FALSE)
print(out$agebin_vector)
print(out$lbin_vector)
print(names(out$agecomp))
print(names(out$lencomp))
}
\author{
Cole Monnahan, Ian Taylor, Sean Anderson, Kelli Johnson
}
\seealso{
\code{\link{sample_lcomp}}, \code{\link{sample_agecomp}}
}

