% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_e.r
\name{change_e_fcast_yrs}
\alias{change_e_fcast_yrs}
\title{Check and change forecast file years if necessary}
\usage{
change_e_fcast_yrs(styr = 0, endyr_orig = 100, endyr_new = 100, fcast_list)
}
\arguments{
\item{styr}{The model start year, an integer}

\item{endyr_orig}{The original end year that the forecast file assumed, an
integer}

\item{endyr_new}{The new end year, an integer}

\item{fcast_list}{forecast file read in using r4ss (is a list)}
}
\value{
A changed forecast list.
}
\description{
Check if forecast years and benchmark years within the forecast file are
within the model start year and end year.
}
