% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_ess.R
\name{change_ESS}
\alias{change_ESS}
\title{Change the effective sample sizes in length and age data in
the EM.}
\usage{
change_ESS(dat_list, ess_len = NULL, ess_ages = NULL)
}
\arguments{
\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}} in the \pkg{r4ss} package. Make sure
you select option \code{section=2}.}

\item{ess_len}{lcomp_params}

\item{ess_ages}{acomp_params}
}
\description{
Change the effective sample sizes in length and age data in
the EM.
}
\details{
reads in the case files information for sample_lencomp,
sample_agecomp or sample_calcomp and changes ESS according to the arguments
in the case files. For now the only option is to use the number of hauls
}
\author{
Gwladys Lambert
}
