% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_p.R
\name{change_p}
\alias{change_p}
\title{Function to modify selectivity parameters in \code{.ctl} files.}
\usage{
change_p(om_dir = "om", om_ctl_file_in = "ss3.ctl",
  om_ctl_file_out = "ss3.ctl", par_file_in = "ss3.par",
  par_file_out = "ss3.par", em_dir = "em", em_ctl_file_in = "ss3.ctl",
  em_ctl_file_out = "ss3.ctl", fleet = c(1, 2), selex_om = c("age",
  "length"), selex_em = c("age", "length"), pars_om_selex = list(c(51, -3,
  5.1, 15, -999, -999), c(51, -3, 5.1, 15, -999, -999)),
  pars_em_selex = list(c(45, -1, 3, 11, -999, -999), c(45, -1, 3, 11, -999,
  -999)), sex_flag = c(1, 1), om_sex_offset = list(c(20, 0, -0.1, -0.1),
  c(20, 0, -0.1, -0.1)), em_sex_offset = list(c(20, 0, -0.1, -0.1), c(20, 0,
  -0.1, -0.1)), plot_selex_om = NULL, plot_selex_em = NULL)
}
\arguments{
\item{om_dir}{Directory to locate SS3 files}

\item{om_ctl_file_in}{Input SS3 control file}

\item{om_ctl_file_out}{Output SS3 control file}

\item{par_file_in}{Input SS3 parameter file}

\item{par_file_out}{Output SS3 parameter file}

\item{em_dir}{Directory to locate SS3 files}

\item{em_ctl_file_in}{Input SS3 control file}

\item{em_ctl_file_out}{Output SS3 control file}

\item{fleet}{Fleet for which the selectivity must be changed}

\item{selex_om}{Type of selectivity. Can take the entry \code{age} or \code{length}}

\item{selex_em}{Type of selectivity. Can take the entry \code{age} or \code{length}}

\item{pars_om_selex}{Selectivity parameters for OM}

\item{pars_em_selex}{Selectivity parameters for EM}

\item{sex_flag}{sex_flag}

\item{om_sex_offset}{Offset selectivity parameters for males in OM}

\item{em_sex_offset}{Offset selectivity parameters for males in EM}

\item{plot_selex_om}{plot selectivity of OM}

\item{plot_selex_em}{plot selectivity of EM}
}
\value{
Altered versions of SS3 \code{.dat}, \code{.ctl} and \code{.par} files are written
to the disk.
}
\description{
Takes ss3 \code{.ctl} files and changes the selectivity parameters.
The function can be called by itself or within \code{\link{run_ss3sim}} to alter 
an operating and estimation model \code{.ctl} files. If used with 
\code{\link{run_ss3sim}} the case file should be named \code{P}. 
A suggested (default) case letter is \code{P} for selex Parameters 
(OK NOT GREAT).
}
\details{
\code{change_p} modifies the selectivity parameters in the \code{.ctl} file.
Selectivity has to be the same type (i.e. either age or length) in both OM and EM. 
The function requires both an OM and EM folder but the same folder path and files 
names can be used if the function is used outside of \code{\link{run_ss3sim}} 
and is used to change the \code{.ctl} file of a single folder. If \code{pars_om_selex} 
is NULL, then thefunction will only change the EM selectivity parameters. If the 
\code{em_selex_shape} is dome shaped the selectivity parameters 5 and 6 will 
be fixed at the value given in \code{pars_em_selex}, the user has to make sure 
he inputs the right parameters that would represent an asymptotic 
selectivity curve.
}
\note{
LOTS OF ROOM FOR IMPROVEMENT I THINK...
}
\section{Which arguments to specifiy in case files}{

All function argument descriptions that start with an asterisk (*) will be passed
through the case files to \code{\link{run_ss3sim}}. If one of these arguments
is not specified in a case file, then a value of \code{NULL} will be passed,
which may or may not be an appropriate value. Other arguments will be ignored
if specified.
}

\examples{
\dontrun{
# Find the example data folders:

temp_path <- tempdir() 
dir.create(temp_path, showWarnings = FALSE)
wd        <- getwd()
setwd(temp_path)

library(ss3models)
d      <- system.file("models", package = "ss3models")
cod    <- paste0(d, "/cod")
file.copy(cod, ".", recursive = TRUE)
setwd("cod")

wd <- getwd()

case_folder <- system.file("extdata", "new-cases", package = "ss3sim")

# With a custom parameters case for selectivity, which we'll call
# the P case. Here, we'll need to define which file the case P should
# read from ("P*-cod.txt"):

# First scenario - 
arg_list <- get_caseargs(case_folder, scenario = "D0-F0-P3-cod",
                         case_files = list(P = "P", D = c("index", "lcomp", "agecomp"), F =
                                             "F"))


### ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~###
### THIS IS THE FUNCTION ALONE - CAN BE RUN FOR ANY SCENARIO       ###
### ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~###

# Here we plot the selectivity curves for each case

library(ss3sim)

setwd(wd)
p_params <- arg_list$P

with(p_params,
     change_p(om_dir               = "om",
              om_ctl_file_in       = "ss3.ctl", ####
              om_ctl_file_out      = "ss3_temp.ctl",
              par_file_in          = "ss3.par",
              par_file_out         = "ss3_temp.par",
              em_dir               = "em",
              em_ctl_file_in       = "ss3.ctl",
              em_ctl_file_out      = "ss3_temp.ctl",
              selex_om               = selex_om,
              selex_em               = selex_em,
              sex_flag            = sex_flag,
              fleet               = fleet, 
              pars_om_selex       = pars_om_selex,
              pars_em_selex       = pars_em_selex,
              om_sex_offset      = om_sex_offset,
              em_sex_offset      = em_sex_offset,
              plot_selex_om       = NULL,
              plot_selex_em       = TRUE
     ))

## NEED TO CLEAN UP AFTER ##
## unlink()
}


}
\seealso{
Other change functions: \code{\link{change_a}},
  \code{\link{change_data}},
  \code{\link{change_em_binning}}, \code{\link{change_e}},
  \code{\link{change_f}}, \code{\link{change_maturity}},
  \code{\link{change_retro}}, \code{\link{change_tv}},
  \code{\link{change_x}}
}
\author{
Gwladys Lambert
}
