% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_q.R
\name{change_q}
\alias{change_q}
\title{Add or remove catchability parameters in a Stock Synthesis control file}
\usage{
change_q(
  string_add = NULL,
  string_remove = NULL,
  ctl_list,
  dat_list,
  ctl_file_in = NULL,
  dat_file_in = NULL,
  ctl_file_out = NULL,
  overwrite = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{string_add}{A vector of character strings with the fleetnames to add.
Numeric values representing the fleet number are also allowed or any
combination thereof.}

\item{string_remove}{A vector of character strings with the fleetnames to remove.
Numeric values representing the fleet number are also allowed or any
combination thereof.}

\item{ctl_list}{A control file read in by \code{\link[r4ss]{SS_readctl}}.}

\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{ctl_file_in}{A string providing the path to the input SS \code{.ctl} file.}

\item{dat_file_in}{A string providing the path to the input SS \code{.dat} file.}

\item{ctl_file_out}{A string providing the path to the output SS control file. If the value is \code{NULL}, the file will not be written to the disk.}

\item{overwrite}{A logical value specifying if the file being written to the
disk should overwrite an existing file with the same name.
The default is to not overwrite files, i.e., \code{overwrite = FALSE}.}

\item{verbose}{When \code{TRUE} messages will be returned
from the function. Often useful for debugging. The default
is \code{FALSE}.}
}
\value{
A modified Stock Synthesis control file list with
the same structure returned by \code{\link[r4ss:SS_readctl]{r4ss::SS_readctl()}}.
}
\description{
Catchability, \eqn{q}, is the proportionality parameter between
fishing effort and population abundance.
That is, when \eqn{q=1},
a survey of the population is thought to encompass the entire population.
Removing or adding a catch-per-unit effort time series
can require changes to the control file because
each catch-per-unit-effort time series requires set up and parameter lines.
}
\details{
In Stock Synthesis, environmental time series can also be modeled as
catch-per-unit-effort time series.
Readers interested in the complete range of functionality should see the
\href{https://nmfs-stock-synthesis.github.io/ss-documentation/SS330_User_Manual.html#catchability}{section of the Stock Synthesis user manual on catchability}.
\verb{code\{change_q\} has limited functionality relative to what is available in Stock Synthesis. For example, }change_q` cannot add a parameter to estimate additional variance.
Though it will remove additional variance parameters for
fleets that no longer have survey data.
Additionally, the float term is not used within ss3sim and is set to zero.

This function can add and remove parameters in the control file simultaneously.
Thus, it is not necessary to call it twice to perform both operations.
}
\seealso{
\code{\link[=check_q]{check_q()}} can determine which fleets should removed or added.
}
\author{
Kelli F. Johnson
}
