% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_rec_devs.r
\name{change_rec_devs}
\alias{change_rec_devs}
\title{Replace recruitment deviations}
\usage{
change_rec_devs(recdevs, ctl_file_in,
  ctl_file_out = "control_recruitment.ss")
}
\arguments{
\item{recdevs}{A vector of recruitment deviations to be entered into
the SS control file. A single value will be repeated for every value in
\code{years} or \code{length(years) == length(recdevs)} must be true.}

\item{ctl_file_in}{A string providing the path to the input SS \code{.ctl} file.}

\item{ctl_file_out}{A string providing the path to the output SS control file. If the value is \code{NULL}, the file will not be written to the disk.}
}
\value{
A modified SS control file.
}
\description{
This function replaces the recruitment deviations in the
control file with those specified in \code{recdevs}.
The new control file is then written to the disk if
\code{ctl_file_out} is specified.
It is imperative that the path provided in \code{ctl_file_in}
be to a \code{ss_new} file so \code{change_rec_devs} can
properly determine how many recruitment deviations to input
into the control file.
}
\details{
This function does not need to be specified in a case file if you
are running an ss3sim simulation using \code{\link{run_ss3sim}}.
}
\examples{
d <- system.file(file.path("extdata", "models"), package = "ss3sim")
change_rec_devs(recdevs = rlnorm(100),
  ctl_file_in = file.path(d, "cod-om", "codOM.ctl"),
  ctl_file_out = file.path(tempdir(), "control_recdevs.ss"))
}
\author{
Kelli Faye Johnson
}
