% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_year.R
\name{change_recyear}
\alias{change_recyear}
\title{Change start year main recruitment deviations in control file}
\usage{
change_recyear(ctl_list, main)
}
\arguments{
\item{ctl_list}{A control file read in by \code{\link[r4ss]{SS_readctl}}.}

\item{main}{An integer specifying the year to
start the main period of recruitment.}
}
\value{
A \code{\link[r4ss:SS_readctl]{r4ss::SS_readctl()}} list with an augmented start year of
the recruitment deviations in the main period.
}
\description{
Change start year main recruitment deviations in control file
}
\author{
Kelli Faye Johnson
}
