% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_year.R
\name{change_startyear}
\alias{change_startyear}
\title{Change start year of the data file}
\usage{
change_startyear(dat_list, firstyear = NULL)
}
\arguments{
\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}} in the \pkg{r4ss} package. Make sure
you select option \code{section=2}.}

\item{firstyear}{An integer value specifying the year in which you want to start
fitting your data. The default is \code{NULL}, which will look to the first year
with a non-zero catch in \code{dat_list} and use that as the first year.}
}
\value{
A \code{\link[r4ss]{SS_readdat}} list with an augmented start year.
}
\description{
Change start year of the data file
}
\author{
Kelli Faye Johnson
}
