% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_tail_compression.R
\name{change_tail_compression}
\alias{change_tail_compression}
\title{Replace tail compression value for length composition data.}
\usage{
change_tail_compression(tail_compression, dat_list, outfile = NULL)
}
\arguments{
\item{tail_compression}{*The new tail_compression value to be used. Must be a
numeric value, as a proportion. For example 0.1 means 10 percent. See the
SS3 manual for further information. A NULL value indicates no action, a
negative value indicates to SS3 to ignore it (not use that feature).}

\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated give all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the 
file takes time and computing resources.}
}
\value{
A modified SS \code{.dat} file, and that file returned invisibly
  (for testing) as a vector of character lines.
}
\description{
This function replaces the tail compression value for length composition data
in a \code{.dat} file that was read in using \code{\link[r4ss]{SS_readdat}}
with those specified in
\code{tail_compression}. It then writes a new file with name \code{dat_file_out}
into the working directory.
}
\author{
Cole Monnahan
}
