% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_x.R
\name{change_x}
\alias{change_x}
\title{Function to mis-specify dimorphism (i.e. remove it) in EM \code{.dat} and \code{.ctl} files.}
\usage{
change_x(dat_file_in, dat_file_out, ctl_file_in, ctl_file_out,
  change_dat = TRUE, change_ctl = TRUE, write_file = TRUE)
}
\arguments{
\item{dat_file_in}{Input SS3 data file}

\item{dat_file_out}{Output SS3 data file}

\item{ctl_file_in}{Input SS3 control file}

\item{ctl_file_out}{Output SS3 control file}

\item{change_dat}{*If TRUE, the function will change the 2 sex model \code{.dat} file
and turn it into single sex}

\item{change_ctl}{*If TRUE, the function will change the 2 sex model \code{.ctl} file
and turn it into single sex}

\item{write_file}{A logical value indicating if the altered file should be
written to the disk. The default value is \code{TRUE}.}
}
\value{
Altered versions of SS3 \code{.dat} and \code{.ctl} files are written
to the disk and the altered \code{dat_file_in} is also returned invisibly.
}
\description{
Takes ss3 \code{.dat} and \code{.ctl} files and removes the sex-specific
parameters to create single sex files. The function can be called by
  itself or within \code{\link{run_ss3sim}} to alter an estimation model
  \code{.dat} and \code{.ctl} file.
  If used with \code{\link{run_ss3sim}} the case file should be named
  \code{X}. A suggested (default) case letter is \code{X} for seX.
}
\details{
\code{change_x} removes sex dimorphism in the \code{.dat} file
by summing up males and females lcomp and agecomp. It also calculates the average 
mlacomp if mlacomp is present in the \code{.dat} file. Growth and selectivity are changed
in the \code{.ctl} file simply by commenting out the male parameters. The initial values 
and bonds should be changed before running the simulation outside of ss3sim. An option 
will be added later on to do that here.
}
\note{
ROOM FOR IMPROVEMENT - MORE CHECKS TO DO... AND COULD DO WITH IMPROVING 
THE ARGUMENTS IN.
}
\section{Which arguments to specifiy in case files}{

All function argument descriptions that start with an asterisk (*) will be passed
through the case files to \code{\link{run_ss3sim}}. If one of these arguments
is not specified in a case file, then a value of \code{NULL} will be passed,
which may or may not be an appropriate value. Other arguments will be ignored
if specified.
}

\examples{
\dontrun{
temp_path <- tempdir()
dir.create(temp_path, showWarnings = FALSE)
wd        <- getwd()
setwd(temp_path)

# Find the SS3sim simple model in the package:
d         <- system.file("extdata/new models", package = "ss3sim")
simple_sex    <- paste0(d, "/GOAflathead")
file.copy(simple_sex, ".", recursive = TRUE)
setwd("GOAflathead")

wd <- getwd()


### ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~###
### THIS IS THE FUNCTION ALONE - CAN BE RUN FOR ANY SCENARIO ABOVE ###
### ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~###

setwd(wd)

out <- change_x(dat_file_in = "om/ss3.dat", dat_file_out = "om/ss3.dat", 
                ctl_file_in = "om/ss3.ctl", ctl_file_out = "om/ss3.ctl", 
                change_dat= TRUE, change_ctl = TRUE,
                write_file=FALSE) # write file controls .dat, not .ctl (always overwrite)

}
}
\seealso{
Other change functions: \code{\link{change_a}},
  \code{\link{change_data}},
  \code{\link{change_em_binning}}, \code{\link{change_e}},
  \code{\link{change_f}}, \code{\link{change_maturity}},
  \code{\link{change_p}}, \code{\link{change_retro}},
  \code{\link{change_tv}}
}
\author{
Gwladys Lambert
}
