% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.r
\name{check_data_str_range}
\alias{check_data_str_range}
\title{Check input arguments for data}
\usage{
check_data_str_range(all_params, dat_list)
}
\arguments{
\item{all_params}{A named list of the parameters containing at a minimum
year and fleet values}

\item{dat_list}{An Stock Synthesis data list object as read in by \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}}.}
}
\description{
Check that the param list inputs have correct structure and range given an
associated data file.
}
