% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_forecast.R
\name{check_forecast}
\alias{check_forecast}
\title{Check input forecast file values}
\usage{
check_forecast(for_list)
}
\arguments{
\item{for_list}{An SS forecast list object as read in from
\code{\link[r4ss]{SS_readforecast}} in the \pkg{r4ss} package.}
}
\value{
A an augmented list object, as returned by \code{\link[r4ss:SS_readforecast]{r4ss::SS_readforecast()}},
is invisibly returned.
}
\description{
Ensure that the \code{forecast.ss} file is configured for use in ss3sim.
}
\details{
\itemize{
\item fish at \eqn{F_{MSY}}
\item use relative benchmark years (i.e., \code{Bmark_years})
\item use relative years for fishing specifications, i.e., \code{Fcast_years}
}
}
\author{
Kelli Faye Johnson
}
