% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clean_data}
\alias{clean_data}
\title{Given sampling arguments, removed ("clean") all data in a .dat file that
is not specified in the arguments.}
\usage{
clean_data(datfile, index_params = NULL, lcomp_params = NULL,
  agecomp_params = NULL, calcomp_params = NULL, mlacomp_params = NULL,
  verbose = FALSE)
}
\arguments{
  \item{index_params,}{lcomp_params, agecomp_params,
  calcomp_params, mlacomp_params Named lists containing the
  arguments for the different sampling functions.}

  \item{verbose}{When \code{TRUE} it will print a message
  when rows are deleted.}
}
\value{
An invisible cleaned data list as an object.
}
\description{
This prepares a .dat file to be used by an EM, whereas
before it may have had leftover data from sampling
purposes. See examples in \code{\link{change_data}}.
}
\note{
This function does not write the result to file.
}
\author{
Cole Monnahan
}
\seealso{
calculate_data_units, change_data
}

