% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod.R
\docType{data}
\name{codemctl}
\alias{codemctl}
\title{Control file for the cod estimation method}
\format{
A list with many items, some of which are highlighted below:
\describe{
\item{fleetnames}{a vector of names for the fleets}
\item{MG_parms}{natural mortality and growth parameters}
\item{SR_parms}{stock-recruitment relationship parameters}
...
}
}
\source{
North Sea cod
(\emph{Gadus morhua}; Richard D. Methot, Jr., NMFS, NOAA, pers. comm.)
}
\usage{
codemctl
}
\description{
A list of controls returned from \code{\link[r4ss:SS_readctl]{r4ss::SS_readctl()}} for the
North Sea cod operating model.
The input file is stored in \code{extdata/models}.
}
\examples{
data("codemctl", package = "ss3sim")
\dontshow{
testthat::expect_true(
  codemctl[["type"]] == "Stock_Synthesis_control_file",
  label = "codemctl is a Stock Synthesis control file"
)
}
}
\seealso{
\itemize{
\item \code{\link[r4ss:SS_readctl]{r4ss::SS_readctl()}}
\item \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}}
}
}
\keyword{datasets}
