% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cod.R
\docType{data}
\name{codomdat}
\alias{codomdat}
\title{Data for the cod operating model}
\format{
A list with many items, some of which are highlighted below:
\describe{
\item{catch}{data frame of catches by year, fleet, and season}
\item{CPUE}{catch-per-unit-effort data}
\item{lencomp}{length-composition data}
...
}
}
\source{
North Sea cod
(\emph{Gadus morhua}; Richard D. Methot, Jr., NMFS, NOAA, pers. comm.)
}
\usage{
codomdat
}
\description{
A list of data returned from \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}} for the
North Sea cod operating model.
The input file is stored in \code{extdata/models}.
}
\seealso{
\code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}}
}
\keyword{datasets}
