% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-results.r
\name{convert_to_wide}
\alias{convert_to_wide}
\title{Convert long-style ss3sim output to wide format}
\usage{
convert_to_wide(lng)
}
\arguments{
\item{lng}{A long dataframe produced from get_results_all().}
}
\value{
A wide dataframe (separate columns for em and om results)
}
\description{
This function exists for back compatibility. Note that this will only work
if the column model_run has only the strings"om" or "em".
}
\examples{
\dontrun{
  scalar <-  read.csv("ss3sim_scalar.csv")
  scalar_wide <- convert_to_wide(scalar)

  ts <- read.csv("ss3sim_ts.csv")
  ts_wide <- convert_to_wide(scalar)
}
}
\author{
Kathryn Doering
}
