% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_ss3models.r
\name{copy_ss3models}
\alias{copy_ss3models}
\title{Copy the OM or EM into a scenario directory}
\usage{
copy_ss3models(model_dir, scenario, iteration = 1, type = c("om", "em"))
}
\arguments{
\item{model_dir}{A directory containing an OM or EM.}

\item{scenario}{A string giving the scenario name which will be used in the
resulting directory name. If you want this directory created somewhere
other than your current working directory, you can pass a full file path
with the last level being the new scenario name. All intermediate
directories that do not exist will be created.}

\item{iteration}{An integer specifying the iteration of interest.}

\item{type}{Either \code{"om"} or \code{"em"} depending on which type of model folder
needs created.}
}
\value{
An invisible boolean for whether that iteration already existed.
}
\description{
Copy the OM or EM into a scenario directory
}
\examples{
# Locate the package data:
om_folder <- system.file(
  "extdata", "models", "cod-om",
  package = "ss3sim"
)

# Copy the operating model:
copy_ss3models(
  model_dir = om_folder,
  scenario = "D0-F0-testing"
)
# Now look at your working directory in your file system

# Copy the EM
copy_ss3models(
  model_dir = om_folder,
  type = "em",
  scenario = "D1-F0-testing"
)
# Scenario argument affects the folder names.

# Clean up:
unlink("D0-F0-testing", recursive = TRUE)
unlink("D1-F0-testing", recursive = TRUE)
}
\author{
Sean C. Anderson, Kelli F. Johnson
}
