% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_em.R
\name{create_em}
\alias{create_em}
\title{Create an EM from an OM}
\usage{
create_em(
  dir_in = system.file("extdata", "models", "cod-om", package = "ss3sim"),
  dir_out = file.path(getwd(), "new-em")
)
}
\arguments{
\item{dir_in}{A file path to a directory that contains the following files:
\code{forecast.ss}, \code{starter.ss}, and a control file
(e.g., \code{xxxOM.ctl}). The default is to get the codOM within \pkg{ss3sim}.}

\item{dir_out}{A file path to a directory where the new files will be saved.
The default is to save the files in your current working directory in a
folder called \code{new-em}.}
}
\value{
Nothing is returned, but three files are saved to the disk in
the specified folder that may also be new.
}
\description{
Create estimation model (EM) files from operating model (OM) files.
By making small changes to the OM rather than having two sets of files,
less files need to be maintained.
Differences between the OM and EM are mainly related to how the
OM takes input fishing mortality values rather than absolute catches.
}
\examples{
create_em()
# The necessary files are in the following folder
dir(file.path(getwd(), "new-em"))
\dontshow{
testthat::expect_true("forecast.ss" \%in\% dir("new-em"),
  label = "forecast file is in 'new-em'"
)
testthat::expect_true("starter.ss" \%in\% dir("new-em"),
  label = "starter file is in 'new-em'"
)
testthat::expect_true("codEM.ctl" \%in\% dir("new-em"),
  label = "codEM file is in 'new-em'"
)
}
# Clean up your directory
unlink(file.path(getwd(), "new-em"), recursive = TRUE)
}
\author{
Kelli Faye Johnson
}
