% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_logo.R
\name{create_logo}
\alias{create_logo}
\title{Create the ss3sim logo}
\usage{
create_logo(outfile = NULL)
}
\arguments{
\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the
file takes time and computing resources.}
}
\value{
A \code{png} file or a graphics device with the
logo used for the \code{ss3sim} project.
}
\description{
Generate and save, if \code{outfile} is provided,
the ss3sim logo using the built-in data.
}
\examples{
ss3sim:::create_logo()
grDevices::dev.off()
}
\author{
Kelli F. Johnson
}
