% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat2simdf.R
\name{dat2dimdf}
\alias{dat2dimdf}
\title{Create a simulation data frame from a data file}
\usage{
dat2dimdf(dat_list)
}
\arguments{
\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}
}
\description{
Create a simulation data frame (\code{simdf}) from a data input file for a
Stock Synthesis operating model. Typically, this data file is not used
to inform the data structure. Here we reverse engineer the ss3sim process
of using \code{simdf} to dictate what the data file looks like and use the data
file to dictate what \code{simdf} looks like.
}
\author{
Kelli F. Johnson
}
