% Generated by roxygen2 (4.0.1): do not edit by hand
\name{facet_form}
\alias{facet_form}
\title{A helper function for building a ggplot facet. Used internally by the
plotting functions.}
\usage{
facet_form(horiz = NULL, horiz2 = NULL, vert = NULL, vert2 = NULL)
}
\value{
A formula which can be used in \code{facet_grid}, or NULL if all
arguments are NULL
}
\description{
A helper function for building a ggplot facet. Used internally by the
plotting functions.
}
\author{
Cole Monnahan
}

