% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fast_SS_writedat.R
\name{fast_SS_writedat}
\alias{fast_SS_writedat}
\title{Writes files faster than SS_writedat with a small modification from the main 
function}
\usage{
fast_SS_writedat(datlist, outfile, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{datlist}{data}

\item{outfile}{out}

\item{overwrite}{overwrite}

\item{verbose}{T or F}
}
\description{
Writes files faster than SS_writedat with a small modification from the main 
function
}
\note{
THIS FUNCTION IS NOW REDUNDANT SINCE THE OPTION WAS ADDED TO THE MAIN 
FUNCTION IN r4ss !!!! NEED TO CHANGE BACK TO SS_writedat WHEREVER THIS HAS 
BEEN USED AND USE THE NEW ARGUMENT: faster = TRUE
}
\author{
Gwladys Lambert
}
