% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_across.r
\name{fill_across}
\alias{fill_across}
\title{Fill in matrix across rows of weight-at-age data by interpolation}
\usage{
fill_across(mat, minYear, maxYear)
}
\arguments{
\item{mat}{A matrix}

\item{minYear}{Minimum year}

\item{maxYear}{Maximum year}
}
\description{
Function that fills in matrix across rows of wtatage data by interpolation
Missing Rows are then backfilled
}
\seealso{
\itemize{
\item \code{\link[=sample_lcomp]{sample_lcomp()}}
\item \code{\link[=sample_agecomp]{sample_agecomp()}}
\item \code{\link[=sample_wtatage]{sample_wtatage()}}
}
}
\author{
Peter Kuriyama and Allan Hicks
}
