% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_across.r
\name{fill_down}
\alias{fill_down}
\title{Fill a matrix using columns.}
\usage{
fill_down(mat)
}
\arguments{
\item{mat}{A numeric matrix, typically information extracted from a
data frame but leaving the meta-data behind. Algebraic functions can
then be applied rows or columns of this matrix.}
}
\description{
Fill a matrix using columns.
}
\examples{
data <- matrix(rnorm(50), ncol = 5)
data[seq(1,50,by = 3)] <- NA
data <- data.frame("Yr" = seq_along(NROW(data)),
"seas" = 1, "sex" = 1, "bio_pattern" = 1,
"birthseas" = 1, "fleet" = 1, data)
colnames(data) <- gsub("^x|^X", "age", colnames(data))
datafill <- fill_down(data)
testthat::expect_equal(datafill[4, 7], c("age1" = mean(data[c(3,5), 7])))
rm(data, datafill)
}
