% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_across.r
\name{fill_header}
\alias{fill_header}
\title{Fill in matrix across rows of weight-at-age data by interpolation}
\usage{
fill_header(data)
}
\arguments{
\item{data}{A matrix}
}
\description{
Function that fills in matrix across rows of wtatage data by interpolation
Missing Rows are then backfilled
}
\seealso{
\code{\link{sample_lcomp}}, \code{\link{sample_agecomp}},
  \code{\link{fill_across}}
}
\author{
Peter Kuriyama and Allan Hicks
}
