\name{find_item}
\alias{find_item}
\title{Find a vector of values from an object between 2 values}
\usage{
  find_item(object, what_vec)
}
\arguments{
  \item{object}{The object (a vector) in which to search}

  \item{what_vec}{A vector of length 2. The function will
  return the values of object between what_vec[1] + 1 and
  what_vec[2] - 2.}
}
\description{
  Find a vector of values from an object between 2 values
}

