\name{fish600-package}
\alias{fish600-package}
\alias{fish600}
\docType{package}
\title{
Functions to manipulate Stock Synthesis input and output files for simulation
}
\description{
This package is based around simulations with underlying time-varying parameters.
}
\details{
\tabular{ll}{
Package: \tab fish600\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2013-03-21\cr
License: \tab MIT\cr
}
Built for a FISH600 project as the University of Washington,  School of Aquatic and Fishery Science
}
\author{
FISH600 class 2013
}
