\docType{data}
\name{fish600seeds}
\alias{fish600seeds}
\title{Randomly generated seeds}
\description{
  These data are randomly generated seed values. They were
  generated with: \code{ set.seed(999) fish600seeds <-
  sample(1:1e6, 100, replace = FALSE) fish600seeds <-
  c(fish600seeds, sample(1:1e6, 900, replace = FALSE)) } I
  have checked to ensure there are no duplicate seeds.
}
\keyword{data}
\keyword{datasets}

