% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case-parsing.r
\name{get_args}
\alias{get_args}
\title{Take a csv file, read it, and turn the first column into the list
names and the second column into the list values.}
\usage{
get_args(file)
}
\arguments{
\item{file}{The file name as character}
}
\description{
Take a csv file, read it, and turn the first column into the list
names and the second column into the list values.
}

