% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/executable.R
\name{get_bin}
\alias{get_bin}
\title{Get SS3 binary/executable location in package}
\usage{
get_bin(filename = "ss3_24o_opt")
}
\arguments{
\item{filename}{Name of SS3 binary (without file extension)}
}
\description{
Get SS3 binary/executable location in package
}
\examples{
get_bin()
}

