\name{get_fish600_casefolder}
\alias{get_fish600_casefolder}
\title{Get the folder location of the FISH600 case files}
\usage{
  get_fish600_casefolder()
}
\value{
  A character object showing the location of the FISH600
  case files in the package \code{extdata} folder.
}
\description{
  Get the folder location of the FISH600 case files
}

