\name{get_fish600_modelfolder}
\alias{get_fish600_modelfolder}
\title{Get the folder location of a FISH600 model}
\usage{
  get_fish600_modelfolder(folder_name)
}
\arguments{
  \item{folder_name}{The model folder name.}
}
\value{
  A character object showing the location of appropriate
  FISH600 folder in the package \code{extdata} folder.
}
\description{
  Get the folder location of a FISH600 model
}
\examples{
\dontrun{
get_fish600_modelfolder("cod-em")
}
}
\author{
  Sean Anderson
}

