% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_comps.R
\name{get_last_phase}
\alias{get_last_phase}
\title{Get the highest phase used in the control file}
\usage{
get_last_phase(ctl)
}
\arguments{
\item{ctl}{A control file list read in using \code{r4ss::SS_readctl}.}
}
\description{
Get the highest phase used in the control file
}
\author{
Kathryn Doering
}
