% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-results.r
\name{get_nll_components}
\alias{get_nll_components}
\title{Get negative log likelihood (NLL) values from a report file list}
\usage{
get_nll_components(report.file)
}
\arguments{
  \item{report.file}{An SS_output list for a model}
}
\description{
Get negative log likelihood (NLL) values from a report file
list
}
\author{
Merrill Rudd
}

