% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-results.r
\name{get_results_derived}
\alias{get_results_derived}
\title{Extract time series from a model run with the associated standard deviation.}
\usage{
get_results_derived(report.file)
}
\arguments{
\item{report.file}{An \code{\link[r4ss]{SS_output}} list for a model
(operating model or estimation model).}
}
\description{
Extract time series from an \code{\link[r4ss:SS_output]{r4ss::SS_output()}} list from a model run.
Returns a data.frame of the results for spawning biomass, recruitment,
forecasts, and effort by year.
}
\seealso{
Other get-results: 
\code{\link{get_results_all}()},
\code{\link{get_results_scalar}()},
\code{\link{get_results_scenario}()},
\code{\link{get_results_timeseries}()}
}
\author{
Kelli F. Johnson
}
\concept{get-results}
