% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-results.r
\name{get_results_mod}
\alias{get_results_mod}
\title{Get results for 1 model run}
\usage{
get_results_mod(dir = getwd(), is_EM = NULL, is_OM = NULL)
}
\arguments{
\item{dir}{The full or relative path to the Stock Synthesis model file folder. If not
specified, uses the working directory.}

\item{is_EM}{Is this an estimation model? Defaults to NULL, which will look
for the letters \code{"EM"} (lower or uppercase) to decide if this is an estimation
model or operating model.}

\item{is_OM}{Is this an operating model? Defaults to NULL, which will look
for the letters \code{"OM"} (lower or uppercase) to decide if this is an estimation
model or operating model.}
}
\value{
A list of 3 data frames called \code{scalar}, \code{timeseries}, and
\code{derived} (for derived quantities). These data frames contain results for 1
model run.
}
\description{
Get results for 1 model run
}
\author{
Kathryn L. Doering
}
