\name{get_results_scalar}
\alias{get_results_scalar}
\title{Extract scalar quantities from a model run.}
\usage{
  get_results_scalar(report.file)
}
\arguments{
  \item{report.file}{An SS_output list for a model (OM or
  EM).}
}
\description{
  Extract scalar quantities from a model run. Returns a
  data.frame of the results (single row) which can be
  rbinded later.
}
\author{
  Cole Monnahan
}

