% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scenarios.R
\name{get_scenarios}
\alias{get_scenarios}
\title{Identify scenarios in \code{directory}}
\usage{
get_scenarios(directory = getwd(), full = FALSE)
}
\arguments{
\item{directory}{The directory or vector of directories that you want to
search for scenarios. The search is recursive, and thus, it is in one's
best interest to provide a shorter path name rather than one high up in the
call stack.}

\item{full}{A logical entry. If \code{TRUE}, the full path name is returned, which
can be helpful if a vector of directories is supplied to \code{directory},
otherwise it is impossible to know which scenarios are located where. If
\code{FALSE}, a vector of names is returned. This is the default behavior.}
}
\value{
A character vector of names of directories that contain output from
{ss3sim}. Full paths are only provided if \code{full = TRUE}.
}
\description{
Find scenario directories, where it is known if a directory was derived from
{ss3sim} and contains iterations of the operating and estimating models if
there are directories with numeric names that contain om and/or em
directories, i.e., "1", "2", "3", ..., "100".
}
\author{
Merrill Rudd
}
\concept{get results}
