% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma.r
\name{get_sigmar}
\alias{get_sigmar}
\title{Get variability about recruitment deviations (\eqn{\sigma_R})}
\usage{
get_sigmar(om)
}
\arguments{
\item{om}{The name of the operating model, which should be the prefix of
the control file, e.g., \code{"myOM"}.
A full directory can be specified with the the prefix of the file name but
leaving off the file extension.}
}
\description{
Use the name of the operating model to open the control file and obtain the
\code{INIT} value for recruitment deviations \eqn{sigma} (\eqn{\sigma_R}).
}
\author{
Kelli F. Johnson
}
