% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss_version.R
\name{get_ss_ver_dl}
\alias{get_ss_ver_dl}
\title{Get ths ss version (either 3.24 or 3.30) from a \code{dat_list} list.}
\usage{
get_ss_ver_dl(dat_list)
}
\arguments{
\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}} in the \pkg{r4ss} package. Make sure
you select option \code{section=2}.}
}
\description{
# Get the SS version from a list \code{dat_list} that was a
  originally read in using \code{\link[r4ss]{SS_readdat}}.
}
