% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss_version.R
\name{get_ss_ver_file}
\alias{get_ss_ver_file}
\title{Get the Stock Synthesis version (either 3.24 or 3.30) from a Stock Synthesis file}
\usage{
get_ss_ver_file(file)
}
\arguments{
\item{file}{Input Stock Synthesis control file, starter, or data file.}
}
\value{
A character value with the version number.
}
\description{
Get the Stock Synthesis version.
This information can be found on the top line in a Stock Synthesis file.
See code in \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}} for an example.
}
