% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_success.R
\name{get_success}
\alias{get_success}
\title{Determine if a Stock Synthesis run was successful}
\usage{
get_success(dir)
}
\arguments{
\item{dir}{A character string specifying the path to the folder with the
results from the run of the Stock Synthesis model.}
}
\value{
A named vector with two values, \code{ran} and \code{hess},
(1) a zero or one for the presence of the \code{Report.sso} file and
(2) a zero or one for the presence of a positive definite hessian matrix.
}
\description{
Use the presence of files generated by Stock Synthesis to determine if the
run was successful or not. There are two levels of success that must
be determined if the run was meant to include estimating the hessian.
}
\author{
Kelli F. Johnson
}
