% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get_vbgf_loglik}
\alias{get_vbgf_loglik}
\title{Function to estimate the log likelihood of a fit to length and age data}
\usage{
get_vbgf_loglik(logL1, logL2, logk, logsigma)
}
\arguments{
  \item{logL1}{numeric, log(L1) parameter.}

  \item{logL2}{numeric, log(L2) parameter.}

  \item{logk}{numeric, log(k) parameter.}

  \item{logsigma}{numeric, log(sigma) parameter, where
  sigma is the slope of the CV line where the CV of an age
  = sigma*age.}

  \item{data_}{data.frame, column 1 is ages and column 2 is
  lengths}

  \item{a3}{numeric, the youngest age well sampled in the
  data.}
}
\description{
Uses \code{\link[bbmle]{mle2}} to minimize the negative log
likelihood and uses \code{\link{vbgf_func}} to predict
data.
}

