% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case-parsing.r
\name{get_weight_comps_args}
\alias{get_weight_comps_args}
\title{return the weight_comps args}
\usage{
get_weight_comps_args(args)
}
\arguments{
\item{args}{List of args created by \code{get_caseargs()}.}
}
\description{
return the weight_comps args
}
