\name{grabSigma}
\alias{grabSigma}
\title{Use the name of the operating model to open the ctl file and obtain the INIT
value for sigmaR (recruitment deviations sigma)}
\usage{
  grabSigma(om)
}
\arguments{
  \item{om}{The name of the operating model, which should
  be the prefix of the .ctl file, eg "myOM"}
}
\description{
  Use the name of the operating model to open the ctl file
  and obtain the INIT value for sigmaR (recruitment
  deviations sigma)
}
\author{
  Kelli Johnson
}

