% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists2df.R
\name{lists2df}
\alias{lists2df}
\title{Combine multiple lists element wise to a single data frame}
\usage{
lists2df(...)
}
\arguments{
\item{...}{Lists of the same length}
}
\value{
A data frame. With one column for each vector present in
the lists, e.g., if the lists have two vectors then the resulting
data frame will have two columns. An additional column will be added
called \code{"index"} that specifies which vector the information
was taken from within each list. For example, if each list has two
vectors and the first vector is of length 3 and the second is of
length 4, then the index column will be \code{c(rep(1,3), rep(2,4))}.
}
\description{
Take the ith element from each list and combine into a data frame
and then rbind the resulting data frames to form a single data frame.
}
\examples{
ss3sim:::lists2df(list(1:2, 5:10), list(3:4, 5:10))
ss3sim:::lists2df(years = list(1:2, 5:10), fvals = list(3:4, 5:10))
ss3sim:::lists2df(years = list(1:2, 5:10))
\dontshow{
testthat::expect_error(
  ss3sim:::lists2df(years = list(1:10, 5:10), fisheries = 1:2)
)
}

}
\author{
Kelli Faye Johnson
}
\keyword{internal}
