% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-results.r
\name{make_df}
\alias{make_df}
\title{Make a list of lists with dataframe components into a dataframes}
\usage{
make_df(list_name, list_df)
}
\arguments{
\item{list_name}{A name to subset from iter_list}

\item{list_df}{A list of dataframes. These need not have the same column
names, as this function will fill in with NAs.}
}
\value{
A dataframe
}
\description{
Bind together list of list components with the same name
}
\author{
Kathryn Doering
}
