% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_re.R
\name{plot_re_panel}
\alias{plot_re_panel}
\title{Plot relative error and indicate median absolute relative error}
\usage{
plot_re_panel(x, re, ylim = c(-0.5, 0.5), mare_pos = max(ylim),
  ypos = c(-0.3, 0, 0.3, 0.6), xaxis = TRUE, yaxis = TRUE,
  col.axis = "grey50", cex.axis = 0.7, cex.mare = 0.7, lty.zero = 1,
  lwd.zero = 0.8, col.zero = "grey80", beans = FALSE, dots = FALSE,
  mare_col_scale = 7, bean_maxwidth = 0.6, col.pts = "#00000030",
  cex.pts = 0.4, jitter.pts = 0.1)
}
\arguments{
  \item{x}{Cases to plot along the x axis}

  \item{re}{A vector of relative errors that aligns with
  the \code{cases} vector.}

  \item{ylim}{Y limits on the plot}

  \item{mare_pos}{Y position to add median absolute
  relative error. Set to \code{NULL} to exclude this.}

  \item{ypos}{Y axis ticks}

  \item{xaxis}{Add an x axis?}

  \item{yaxis}{Add a y axis?}
}
\description{
Plot relative error and indicate median absolute relative
error
}
\examples{
set.seed(1)
x <- data.frame(case = c(rep(c(1, 2), each = 10)),
  re = c(rnorm(10, sd = 0.3), rnorm(10, 0.02, sd = 0.2)))
plot_re_panel(x$case, x$re, beans = TRUE, ylim = c(-1, 1), dots = TRUE, box = TRUE)
}

