% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile_fmsy.r
\name{profile_fmsy}
\alias{profile_fmsy}
\title{Determine Fmsy for a given operating model}
\usage{
profile_fmsy(
  om_in,
  results_out,
  start = 0,
  end = 1.5,
  by_val = 0.01,
  verbose = FALSE
)
}
\arguments{
\item{om_in}{A full or relative path to a directory that contains an
\pkg{ss3sim} operating model.}

\item{results_out}{A full or relative path to a directory where the results
will be saved. The directory will be created if it does not already exist.}

\item{start, end}{A single numerical value for each argument specifying the
lowest and highest fishing levels that you want to explore for the fishing
fleet in your model.}

\item{by_val}{Interval at which fishing mortality will be incremented
between \code{start} and \code{end} using \code{seq(start, end, by_val)}.}

\item{verbose}{When \code{TRUE} messages will be returned
from the function. Often useful for debugging. The default
is \code{FALSE}.}
}
\value{
A data frame of catch by fishing mortality is returned invisibly and
saved to the disk along with a figure, \code{Fmsy.pdf}.
}
\description{
Runs an operating model over a range of fishing mortality, \eqn{F}, levels to
determine the F at maximum sustainable yield, \eqn{F_{MSY}}.
}
\details{
\code{profile_fmsy()} runs the operating model with a constant level of
fishing for each year and extracts the expected catch in the terminal year.
It is assumed that the model time series is long enough for the population
to come to equilibrium, and thus, the catch in the terminal year is
equivalent to equilibrium catch.

If the function is run with \code{verbose = TRUE}, which is not the default,
the coefficient of variations of the catches in the terminal years of
the model will be printed to the screen.
Here, terminal is defined as half as many years as there are
ages in the population dynamics of your model.
Thus, if the population plus group starts at age twenty,
the standard deviation of the last ten years of catch
divided by the mean catch over that same time will be printed to the
screen for each model that is ran. For the default cod model provided within
the package, the CV is less than 1e-04 for all explored levels of fishing
mortality.

Ensure that the argument \code{om_in} leads to an operating model that is
configured for use within \pkg{ss3sim}. For example, the \eqn{F} type must
allow for an input vector of \eqn{Fs} rather than catches, along with other
specifications.
}
\examples{
\dontrun{
d <- system.file("extdata", "models", "cod-om", package = "ss3sim")
fmsy.val <- profile_fmsy(
  om_in = d, results_out = "fmsy",
  start = 0.1, end = 0.2, by_val = 0.05
)
# cleanup
unlink("fmsy", recursive = TRUE)
}

}
