\docType{data}
\name{recdevs}
\alias{recdevs}
\title{Randomly generated recruitment deviations}
\description{
  These data are randomly generated recruitment deviations
  with mean 0 and standard deviation 1. They were generated
  with \code{set.seed(999);recdevs <- matrix(nrow = 100,
  ncol = 100, data = rnorm(100*100)}
}
\keyword{data}
\keyword{datasets}

