% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_q.R
\name{remove_CPUE}
\alias{remove_CPUE}
\title{Remove a q setup line into an SS control file}
\usage{
remove_CPUE(string, ctl.in, ctl.out, dat.in, dat.out, overwrite = FALSE)
}
\arguments{
\item{string}{A string with the fleetname to remove.}

\item{ctl.in}{An SS control file name to read in.}

\item{ctl.out}{The SS control file to read out.}

\item{dat.in}{An SS data file name to read in.}

\item{dat.out}{An SS data file name to read out.}

\item{overwrite}{Logical. Overwrite an existing file with the same name as
\code{ctl.out} or \code{data.out}?}
}
\value{
A modified SS control file.
}
\description{
This function removesa q setup line from a SS 3.30 control file
}
\author{
Kelli Johnson
}
