% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_q.R
\name{remove_q_ctl}
\alias{remove_q_ctl}
\title{Remove a q setup line into an SS control file only}
\usage{
remove_q_ctl(string, ctl.in, filename = TRUE, ctl.out,
  overwrite = FALSE)
}
\arguments{
\item{string}{A string with the fleetname to remove.}

\item{ctl.in}{An SS control file name to read in if filename = TRUE otherwise
an SS control file vector already read using readLines()}

\item{filename}{Does function expect ctl.in to be a filename? defaults to
TRUE. See \code{ctl.in} definition.}

\item{ctl.out}{The SS control file to read out.}

\item{overwrite}{Logical. Overwrite an existing file with the same name as
\code{ctl.out} or \code{data.out}?}
}
\value{
A modified SS control file vector.
}
\description{
This function removesa q setup line from a SS 3.30 control file
}
\author{
Kelli Johnson
}
