% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_x.R
\name{replace_x}
\alias{replace_x}
\title{Replace a \code{NULL} value with \code{NA} in a list}
\usage{
replace_x(x, replacement = NA_integer_)
}
\arguments{
\item{x}{An object that potentially has a length of zero and you
wish it to be an actual value.}

\item{replacement}{The value you would like to use to replace items
with a length of zero. For example, the default \code{NA_integer_}
will replace all \code{NULL} values with {NA}. Other options for this
argument could be \code{NA_character_}.}
}
\value{
The object \code{x} is returned with some items replaced.
If the input object was of zero length, then the \code{replacement}
parameter will  be returned instead.
}
\description{
Replace items with a zero length in a list with the value supplied in
the argument \code{replacement}. Useful for scenarios where \code{NULL}
is sometimes a valid input from legacy code. Nested  lists can behave
badly when they have a \code{NULL} entry, for example when converting
to a \code{\link[tibble:tibble]{tibble::tibble()}}, it will be unnamed.
}
\examples{
employees <- list(
  list(
    id = 1,
    dept = "IT",
    age = 29,
    sportsteam = "softball"
  ),
  list(
    id = 2,
    dept = "IT",
    age = 30,
    sportsteam = NULL
  ),
  list(
    id = 3,
    dept = "IT",
    age = 29,
    sportsteam = "hockey"
  ),
  list(
    id = 4,
    dept = NULL,
    age = 29,
    sportsteam = "softball"
  )
)
# Meat of the example here!
\dontrun{
do.call(rbind, lapply(employees, rbind)) |>
  data.frame() |>
  purrr::modify_depth(2, replace_x)
}

}
\author{
Amanda from \href{https://stackoverflow.com/questions/22870198/is-there-a-more-efficient-way-to-replace-null-with-na-in-a-list}{stack overflow}
}
