\name{run_model}
\alias{run_model}
\title{Run the operating models for a specified set of scenario IDs}
\usage{
  run_model(scenarios, iterations, type = c("om", "em"))
}
\arguments{
  \item{scenarios}{Which scenarios to run.}

  \item{iterations}{Which iterations to run.}

  \item{type}{Are you running the operating or estimation
  models?}
}
\description{
  Run the operating models for a specified set of scenario
  IDs
}
\author{
  Sean Anderson
}

