\name{run_om}
\alias{run_om}
\title{Run the operating models for a specified set of scenario IDs}
\usage{
  run_om(scenarios, iterations)
}
\arguments{
  \item{scenarios}{Which scenarios to run.}

  \item{iterations}{Which iterations to run.}
}
\description{
  Run the operating models for a specified set of scenario
  IDs
}
\author{
  Sean Anderson
}

