% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_catch.R
\name{sample_catch}
\alias{sample_catch}
\title{Sample the catches with observation error}
\usage{
sample_catch(dat_list, outfile = NULL)
}
\arguments{
\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the
file takes time and computing resources.}
}
\value{
A modified \code{.dat} file if \code{!is.null(outfile)}. A list object
containing the modified \code{.dat} file is returned invisibly.
}
\description{
This function creates a matrix of sampled catches from the expected
available catches for all fleets with catches.
The input value used for \code{catch_se} will be used to resample the catches.
There is a bit of a disconnect here because catches are defined by input \code{F}
values not absolute catches.
Let \eqn{D_y} be the discard
from the operating model for year y. Then the sampled value is calculated as:
\eqn{D_y*exp(rnorm(1, 0, sds_obs)-sds_obs^2/2)}. The second term
adjusts the random samples so that their expected value is \eqn{D_y}, i.e.,
the log-normal bias correction.
}
\seealso{
Other sampling functions: 
\code{\link{clean_data}()},
\code{\link{sample_agecomp}()},
\code{\link{sample_calcomp}()},
\code{\link{sample_discard}()},
\code{\link{sample_index}()},
\code{\link{sample_lcomp}()},
\code{\link{sample_mlacomp}()},
\code{\link{sample_wtatage}()}
}
\author{
Kelli F. Johnson
}
\concept{sampling functions}
