% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample_ccomp}
\alias{sample_ccomp}
\title{Sample conditional age-at-length compositions from expected values}
\usage{
sample_ccomp(infile, outfile, fleets = c(1, 2), years, yearxbin, cpar = 1,
  write_file = TRUE, keep_agecomp = TRUE)
}
\arguments{
  \item{infile}{An SS data object as read in from
  \code{\link[r4ss]{SS_readdat}} in the \pkg{r4ss} package.
  Make sure you select option \code{section=2}.}

  \item{outfile}{A character string of the new \code{.dat}
  file name to be created. Must end in \code{.dat}.}

  \item{fleets}{*A numeric vector giving the fleets to be
  used. This order also pertains to other arguments. A
  missing value excludes that fleet from \code{outfile}
  (i.e. it turns it off so no samples are written). If none
  of the fleet collected samples, keep the value to
  \code{fleets=NULL}.}

  \item{years}{*A list the same length as \code{fleets}
  giving the years as numeric vectors. If no fleet
  collected samples, keep the value to \code{years=NULL}.}

  \item{yearxbin}{A list of lists indicating the length
  bins with samples for each fleet. Desired bins for each
  year of included data should be available for each fleet.
  If bins do not change between years a single value for
  each fleet can be specified in the argument.}

  \item{write_file}{A switch for whether to write
  \code{outfile} to disk. Can be turned off to speed up
  testing or exploration of the function. The new data are
  returned invisibly, as in the examples below.}

  \item{keep_agecomp}{A logical if age compositions data
  should be kept or removed entirely from the \code{.dat}
  file.  \code{sample_ccomp} only works on the conditional
  age-at-length data and not on the age composition data.
  To sample the composition data set \code{keep_agecomp} to
  \code{TRUE} and use \link{\code{sample_agecomp}}.}
}
\value{
A modified \code{.dat} file if \code{write_file=TRUE}. A
list object containing the modified \code{.dat} file is
returned invisibly.
}
\description{
Take a \code{data.SS_new} file containing expected values
and sample to create observed conditional length-at-age
compositions which are then written to file for use by the
estimation model.
}
\examples{
d <- system.file("extdata", package = "ss3sim")
f_in <- paste0(d, "/example-om/data.ss_new")
infile <- r4ss::SS_readdat(f_in, section = 2, verbose = FALSE)

## Turn off age comps by specifying fleets=NULL
sample_agecomp(infile=infile, outfile="test1.dat",
               fleets=NULL, cpar=c(5,NA), Nsamp=list(100,100),
               years=list(1995, 1995), write_file=FALSE)

## Generate with a smaller number of fleet taking samples
ex1 <- sample_agecomp(infile=infile, outfile="test1.dat", fleets=c(2),
                      Nsamp=list(c(10,50)), years=list(c(1999,2000)),
                      write_file=FALSE)

## Generate with varying Nsamp by year for first fleet
ex2 <- sample_agecomp(infile=infile, outfile="test2.dat", fleets=c(1,2),
                      Nsamp=list(c(rep(50, 5), rep(100, 5)), 50),
                      years=list(seq(1994, 2012, by=2),
                          2003:2012), write_file=FALSE)

## Run three  cases showing Multinomial, Dirichlet(1) and over-dispersed
## Dirichlet for different levels of sample sizes
op <- par(mfrow = c(1,3))
for(samplesize in c(30, 100, 1000)){
    ex4 <- sample_agecomp(infile=infile, outfile="test4.dat", fleets=c(1,2),
                          Nsamp=list(samplesize, samplesize),
                          write_file = FALSE,
                          years=list(2000,2000), cpar=c(NA, 1))
    ex5 <- sample_agecomp(infile=infile, outfile="test5.dat", fleets=c(1,2),
                          Nsamp=list(samplesize, samplesize),
                          write_file = FALSE,
                          years=list(2000,2000), cpar=c(1, 1))
    ex6 <- sample_agecomp(infile=infile, outfile="test6.dat", fleets=c(1,2),
                          Nsamp=list(samplesize, samplesize),
                          write_file = FALSE,
                          years=list(2000,2000), cpar=c(5, 1))
    true <- subset(infile$agecomp, FltSvy==1 & Yr == 2000)[-(1:9)]
    true <- true/sum(true)
    plot(0:15, subset(ex4, FltSvy==1)[1,-(1:9)], type="b", ylim=c(0,1),
         col=1, xlab="Age", ylab="Proportion", main=paste("Sample size=",
         samplesize))
    legend("topright", legend=c("Multinomial", "Dirichlet(1)",
                                "Dirichlet(5)", "Truth"),
           lty=1, col=1:4)
    lines((0:15), subset(ex5, FltSvy==1)[1,-(1:9)], type="b", col=2)
    lines((0:15), subset(ex6, FltSvy==1)[1,-(1:9)], type="b", col=3)
    lines((0:15), true, col=4, lwd=2)
}
par(op)
}
\author{
Kelli Faye Johnson
}
\seealso{
\code{\link{sample_lcomp}} \code{\link{sample_agecomp}}
}

