% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_discard.r
\name{sample_discard}
\alias{sample_discard}
\title{Sample the discard with observation error}
\usage{
sample_discard(
  dat_list,
  outfile = NULL,
  fleets,
  years,
  sds_obs,
  seas = list(1)
)
}
\arguments{
\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated give all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the 
file takes time and computing resources.}

\item{fleets}{*A vector of integers specifying which fleets to include.
The order of the fleets pertains to the input order of other arguments.
An entry of \code{fleets=NULL} leads to zero samples for any fleet.}

\item{years}{*A list the same length as \code{fleets} giving the years as
numeric vectors. If no fleet collected samples, keep the value to
\code{years=NULL}.}

\item{sds_obs}{A list the same length as \code{fleets}. The list should
contain either single values or numeric vectors of the same length as the
number of years which represent the standard deviation of the observation
error. Single values are repeated for all years.}

\item{seas}{A list the same length as \code{fleets}. The list should contain
either single numeric values or numeric vectors, where the length of each
vector matches the length of each vector present in \code{years}.
Single values are repeated for all years.
This parameter operates the same as \code{sds_obs} but it specifies the
season rather than the standard deviation of the observations.}
}
\value{
A modified \code{.dat} file if \code{!is.null(outfile)}. A list object
containing the modified \code{.dat} file is returned invisibly.
}
\description{
This function creates an index of discards sampled from the expected
available discards for specified fleets in specified years. Let \code{D_y} be the discard
from the operating model for year y. Then the sampled value is calculated as:
\code{D_y*exp(rnorm(1, 0, sds_obs)-sds_obs^2/2)}. The second term
adjusts the random samples so that their expected value is \code{D_y}
(i.e., the log-normal bias correction).
}
\seealso{
Other sampling functions: 
\code{\link{clean_data}()},
\code{\link{sample_agecomp}()},
\code{\link{sample_calcomp}()},
\code{\link{sample_index}()},
\code{\link{sample_lcomp}()},
\code{\link{sample_mlacomp}()},
\code{\link{sample_wtatage}()}
}
\author{
Kelli F. Johnson
}
\concept{sampling functions}
