% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample_fit_vbgf}
\alias{sample_fit_vbgf}
\title{Function to sample the log likelihood of a fit to length and age data}
\usage{
sample_fit_vbgf(length.data, start.L1, start.L2, start.k, start.cv.young,
  start.cv.old, a3, A)
}
\arguments{
  \item{length.data}{data.frame which contains the lengths
  and ages to fit the vbgf model..}

  \item{start.L1}{numeric, starting guess for mle2 for L1
  parameter.}

  \item{start.L2}{numeric, starting guess for mle2 for L2
  parameter.}

  \item{start.k}{numeric, starting guess for mle2 for k
  parameter.}

  \item{start.cv.young,}{starting guess for mle2 for
  cv.young parameter.}

  \item{a3}{integer, the youngest age well sampled in the
  data.}

  \item{A}{integer, the oldest age well sampled in the
  data.}
}
\description{
Function to sample the log likelihood of a fit to length
and age data
}

