% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sample_mlacomp}
\alias{sample_mlacomp}
\title{Sample mean length (size-)-at-age data and write to file for use by the EM.}
\usage{
sample_mlacomp(datfile, outfile, ctlfile, fleets = 1, years,
  write_file = TRUE)
}
\arguments{
  \item{datfile}{A path to the data file, outputed from an
  OM, containing the true age distributions (population
  bins). This file is read in and then used to determine
  how many fish of each age bin are to be sampled.}

  \item{outfile}{A character string of the new \code{.dat}
  file name to be created. Must end in \code{.dat}.}

  \item{ctlfile}{A path to the control file, outputed from
  an OM, containing the OM parameters for growth and
  weight/length relationship. These values are used to
  determine the uncertainty about weight for fish sampled
  in each age bin.}

  \item{fleets}{*A numeric vector giving the fleets to be
  used. This order also pertains to other arguments. A
  missing value excludes that fleet from \code{outfile}
  (i.e. it turns it off so no samples are written). If none
  of the fleet collected samples, keep the value to
  \code{fleets=NULL}.}

  \item{years}{*A list the same length as \code{fleets}
  giving the years as numeric vectors. If no fleet
  collected samples, keep the value to \code{years=NULL}.}

  \item{write_file}{A switch for whether to write
  \code{outfile} to disk. Can be turned off to speed up
  testing or exploration of the function. The new data are
  returned invisibly, as in the examples below.}

  \item{infile}{An SS data object as read in from
  \code{\link[r4ss]{SS_readdat}} in the \pkg{r4ss} package.
  Make sure you select option \code{section=2}.}

  \item{Nsamp}{*A numeric list of the same length as
  \code{fleets}. Either single values or vectors of the
  same length as the number of years can be passed through.
  Single values are repeated for all years. If no fleet
  collected samples, keep the value to \code{Nsamp=NULL}.}

  \item{cpar}{*A numeric value or vector the same length as
  \code{fleets} controlling the variance of the Dirichlet
  distribution used for sampling. A value of \code{1}
  indicates the same standard deviation as a multinomial of
  the given \code{Nsamp}, \code{2} indicates twice, etc.
  Values greater than one indicate overdispersion, and less
  underdispersion.}
}
\value{
A modified \code{.dat} file if \code{write_file=TRUE}. A
list object containing the modified \code{.dat} file is
returned invisibly.
}
\description{
Sample mean length (size-)-at-age data and write to file
for use by the EM.
}
\details{
Take a \code{data.SS_new} file containing expected values
and sample from true ages to get realistic proportions for
the number of fish in each age bin, then use the mean
size-at-age and CV for growth to generate random samples of
size, which are then averaged to get mean weight-at-age
values. These values are then written to file for the EM.
}
\section{Which arguments to specifiy in case files}{
  All function argument descriptions that start with an
  asterisk (*) will be passed through the case files to
  \code{\link{run_ss3sim}}. If one of these arguments is
  not specified in a case file, then a value of \code{NULL}
  will be passed, which may or may not be an appropriate
  value. Other arguments will be ignored if specified.
}
\author{
Cole Monnahan
}
\seealso{
\code{\link{sample_lcomp}, \link{sample_agecomp}}
}

