% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mn.R
\name{sample_mn}
\alias{sample_mn}
\title{Sample with a multinomial distribution}
\usage{
sample_mn(data, n)
}
\arguments{
\item{data}{A data frame with one row.}

\item{n}{The desired sample size.}
}
\value{
A data frame with one row because right now the input
data should only be a single row of data.
}
\description{
Sample with a multinomial distribution
}
