% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_wtatage.R
\name{sample_wtatage}
\alias{sample_wtatage}
\title{Sample empirical weight-at-age data and write to file for use by the EM}
\usage{
sample_wtatage(
  outfile,
  dat_list,
  ctl_file_in,
  years,
  fill_fnc = fill_across,
  fleets,
  cv_wtatage = NULL,
  distribution = c("rmultinom")
)
}
\arguments{
\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the 
file takes time and computing resources.}

\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}} in the \pkg{r4ss} package. Make sure
you select option \code{section=2}.}

\item{ctl_file_in}{A path to the control file, output from an OM, containing
the OM parameters for growth and weight/length relationship. These values
are used to determine the uncertainty about weight for fish sampled in each
age bin. Commonly \code{control.ss_new}}

\item{years}{*A list the same length as \code{fleets} giving the years as
numeric vectors. If no fleet collected samples, keep the value to
\code{years=NULL}.}

\item{fill_fnc}{*A function to fill in missing values (ages and years). The
resulting weight-at-age file will have values for all years and ages.One
function is \code{fill_across}.}

\item{fleets}{*A vector of integers specifying which fleets to include.
The order of the fleets pertains to the input order of other arguments.
An entry of \code{fleets=NULL} leads to zero samples for any fleet.}

\item{cv_wtatage}{A user specified CV for growth. Default is \code{NULL}.}

\item{distribution}{Specify the distribution to sample from. Currently, only one distribution
is available, the multinomial.}
}
\value{
A modified \code{.wtatage.ss} file if \code{!is.null(outfile)}. A list
  object containing the modified \code{.wtatage.ss} file is returned invisibly.
}
\description{
Take a \code{data.SS_new} file containing expected values and sample from
true ages to get realistic proportions for the number of fish in each age
bin, then use the mean size-at-age and CV for growth to generate random
samples of size, which are then converted to weight and averaged to get mean
weight-at-age values. Missing ages and years are filled according to a
specified function. These matrices are then written to file for the EM. By
calling this function, \pkg{ss3sim} will turn on the empirical weight-at-age
function (set maturity option to 5) automatically. See
\code{\link{ss3sim_base}} for more details on how that is implemented.
If used with \code{\link{run_ss3sim}} the case file should be named
\code{wtatage}.
}
\section{Which arguments to specifiy in case files}{

All function argument descriptions that start with an asterisk (*) will be passed
through the case files to \code{\link{run_ss3sim}}. If one of these arguments
is not specified in a case file, then a value of \code{NULL} will be passed,
which may or may not be an appropriate value. Other arguments will be ignored
if specified.
}

\examples{
d <- system.file("extdata", "models", "cod-om", package = "ss3sim")
dat <- r4ss::SS_readdat(dir(d, full.names = TRUE, pattern = "\\\\.dat"),
  verbose = FALSE)
dat[["use_MeanSize_at_Age_obs"]] <- 1
temp <- t(sapply(rnorm(length(26:100), mean = 10), 
  seq, to = 100, length.out = dat$Nages))
colnames(temp) <- paste0("a", 1:dat$Nages)
dat[["MeanSize_at_Age_obs"]] <- data.frame(
  "Yr" = 26:100, "Seas" = 1, "FltSvy" = 1, "Gender" = 0, "Part" = 0,
  "AgeErr" = 0, "Ignore" = 10, temp)
wt <- sample_wtatage(outfile = NULL, dat_list = dat,
 ctl_file_in = dir(d, full.names = TRUE, pattern = "\\\\.ctl"),
 years = list(50:100), fill_fnc = fill_across,
  fleets = list(1), cv_wtatage = 0.05)

rm(d, dat, temp, wt)

}
\seealso{
\code{\link{fill_across}}

Other sampling functions: 
\code{\link{clean_data}()},
\code{\link{sample_agecomp}()},
\code{\link{sample_calcomp}()},
\code{\link{sample_index}()},
\code{\link{sample_lcomp}()},
\code{\link{sample_mlacomp}()}
}
\author{
Cole Monnahan, Allan Hicks, Peter Kuriyama
}
\concept{sampling functions}
