% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ss3model.r
\name{sanitize_admb_options}
\alias{sanitize_admb_options}
\title{Check ADMB options to make sure there aren't flags there shouldn't be}
\usage{
sanitize_admb_options(x, exclude = "-nohess")
}
\arguments{
\item{x}{The ADMB options.}

\item{exclude}{A string (not a vector).}
}
\description{
Check ADMB options to make sure there aren't flags there shouldn't be
}
\author{
Sean C. Anderson
}
