% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists2df.R
\name{scalar2list}
\alias{scalar2list}
\title{Replicate an object a number of times to return a list}
\usage{
scalar2list(scalar, length)
}
\arguments{
\item{scalar}{A single numeric value.}

\item{length}{A vector of integer values specifying how many times \code{scalar}
should be repeated in each list element.}
}
\value{
A list of the same length as \code{length}. If, \code{scalar} is not
a list but a vector, then it will be returned as a list.
}
\description{
Replicate the input scalar argument the provided number of times,
where \code{length} can be a vector, and thus, the returned object
will be a list with potentially more than one entry in the list.
This is helpful when an input argument needs to be repeated to match
other input arguments that are lists of vectors when creating data frames.
}
\keyword{internal}
