% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scalar_dat.r
\docType{data}
\name{scalar_dat}
\alias{scalar_dat}
\title{Example scalar data from the Introduction vignette}
\description{
An R object read in using \code{read.csv("ss3sim_scalar.csv")} after
processing the results from the Introduction vignette
using \code{\link[=get_results_all]{get_results_all()}}. The data set is available
so users do not have to wait for the scenarios to run.
}
\examples{
data("scalar_dat", package = "ss3sim")

\dontshow{
testthat::expect_true("max_grad" \%in\% colnames(scalar_dat),
  label = "the column 'max_grad' of the scalar data set")
}

}
\keyword{data}
