\docType{data}
\name{scalars}
\alias{scalars}
\title{Example scalar data output}
\description{
  These data represent the output from a simple example in
  which the observation error on a survey is manipulated.
  Case \code{D0} refers to the base case and case \code{D1}
  refers to a case in which the standard deviation on the
  survey observation error is halved.
}
\keyword{data}
\keyword{datasets}

