% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schooling_pattern.R
\name{schooling_pattern}
\alias{schooling_pattern}
\title{Draws an aggregation pattern from empirical data}
\usage{
schooling_pattern(mean_lengths, SD, new_lengths, plot_fit = T)
}
\arguments{
\item{mean_lengths}{vector of observed mean lengths (i.e. mean length in haul)}

\item{SD}{vector of associated standard deviations (i.e. deviations in length
in haul)}

\item{new_lengths}{vector of lengths for which to generate SDs.}

\item{plot_fit}{creates 2 plots. One that show the empirical relationship between 
mean and SD, overlaid with GAM confidence interval, and another one with GAM 
predictions, prediction interval and new SDs.}
}
\value{
A dataset of \code{new_lengths} and generated SDs and corresponding 
Gamma distribution parameters
}
\description{
Takes a dataset of mean lengths and associated standard deviations,
that should be empirical data, i.e. mean length and SD from a number of hauls,
fits a Gamma GAM model and draws new SD predictions for a given
vector of mean lengths. These mean and SDs are the parameters to be used to 
define "clusters". However those clusters can not be normally ditributed as 
they are used to sample lengths data (ie positive) so the function also 
provides the equivalent Gamma distribution parameters for each mean length/SD 
combination
}
\details{
The function can be used independantly but was specifically written 
to be integrated into the length sampling function. See
 \code{\link{sample_lcomp}}. It is advised to use the \code{parallel_lgths}
 when turning the schooling option on in \code{\link{sample_lcomp}}.
}
\examples{
 \dontrun{
 mean_lengths <- seq(1:500)
 SD <- 0.1*1:500 + rnorm(100, 0, 5) # SHOULD PUT BETTER EXAMPLE - 
                                    # E.G. DOME SHAPE
 new_lengths = mean_lengths
 schooling_pattern(mean_lengths, SD, new_lengths, plot_fit=T) 
 }
 
}
\seealso{
Other sampling functions: \code{\link{clean_data}},
  \code{\link{sample_agecomp}},
  \code{\link{sample_calcomp}}, \code{\link{sample_index}},
  \code{\link{sample_lcomp}}, \code{\link{sample_mlacomp}},
  \code{\link{sample_schools}},
  \code{\link{sample_wtatage}}
}
\author{
Gwladys Lambert
}
