% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_bins.R
\name{setup_bins}
\alias{setup_bins}
\title{Set up bin structure for composition data}
\usage{
setup_bins(bins, nsex = 1, leader = c("l", "a"))
}
\arguments{
\item{bins}{A vector of integer values, either lengths
or ages. Do not repeat them if you are using a two-sex model,
the function will do that for you.}

\item{nsex}{A single integer of one or two specifying the number
of sexes in the model.}

\item{leader}{Most users will not need to change the leader
character from the default unless you are working with age data,
then just use \\code{"a"} rather than the default of \code{"l"}.
If you have a two-sex model, i.e., \code{nsex = 2}, then the
function will take care of the naming for you, which is forced
to \code{"f"} and \code{"m"}. You can change them afterwards if you
want.}
}
\description{
Set up the bin structure needed for composition data.
}
\examples{
ex <- setup_bins(bins = 1:10, nsex = 2, leader = "a")
test <- length(ex) == 20 & all(grep("m", ex) == 11:20)
\dontshow{
testthat::expect_true(test)
}
ex <- setup_bins(bins = 1:5, nsex = 1)
test <- ex[4] == "l4"
\dontshow{
testthat::expect_true(test)
}
}
\seealso{
This is a helper function used to create the bins before
sampling takes place, see \code{\link[=ss3sim_base]{ss3sim_base()}}.
}
