% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_scenarios.R
\name{setup_scenarios}
\alias{setup_scenarios}
\title{Set up scenarios for a simulation with ss3sim}
\usage{
setup_scenarios(input = NULL)
}
\arguments{
\item{input}{A data frame with one row per scenario. If \code{NULL},
which is the default, then a generic scenario will be run with minimal
data to check that everything is working.}
}
\description{
Set up scenarios with default arguments for a simulation with ss3sim.
The data frame passed by the user will likely be a truncated version
of all of the information needed to run a simulation, and therefore,
additional arguments will be added at their default values. The resulting
data frame is then turned into a list and passed to \code{\link{ss3sim_base}}
within \code{\link{run_ss3sim}}.
}
\author{
Kelli Faye Johnson
}
