% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_scenarios.R
\name{setup_scenarios_lookup}
\alias{setup_scenarios_lookup}
\title{Create a named vector to look up full names for types of arguments}
\usage{
setup_scenarios_lookup()
}
\description{
Create a named vector to look up full names for types of arguments
}
