% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_scenarios.R
\name{setup_scenarios_name}
\alias{setup_scenarios_name}
\title{Create a name for an unnamed scenario}
\usage{
setup_scenarios_name(check = FALSE)
}
\arguments{
\item{check}{A logical that enables checking for a unique name.
If \code{check = TRUE} then the function enters a loop and will generate
a names until it finds one that doesn't already exist.
This could be helpful when running scenarios in parallel.}
}
\value{
A single character value is returned.
The object starts with the letter \code{s} and is followed by \link{Sys.time}
Where, the date/time portion is \verb{\%m\%d\%H\%M\%S}, better known as
a two-digit month, e.g., 01; a two-digit number for the day of the month;
and finally a two-digit hour, then minute, then second.
}
\description{
Create a name for an unnamed scenario based on \link{Sys.time}.
}
