% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_scenarios.R
\name{setup_scenarios_rm_NAs}
\alias{setup_scenarios_rm_NAs}
\title{Remove Fleets with NA Input Arguments}
\usage{
setup_scenarios_rm_NAs(scen_params)
}
\arguments{
\item{scen_params}{A list of parameters for at least one scenario, where
each scenario will be a list object of the main list to allow for the
use of \code{\link{lapply}}.}
}
\description{
Remove fleets from the settings list for a given scenario that have one
input argument as NA suggesting that sampling shouldn't happen for that
fleet. NAs from the data frame are moved to a list form with one object
per scenario, and this function removes those NA fleets from the list.
}
