% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_comp.R
\name{split_comp}
\alias{split_comp}
\title{Break apart composition data}
\usage{
split_comp(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A list of four data frames and two integer values. The latter specify
the number of columns that contain meta data and the number of rows of data.
}
\description{
Composition data typically includes meta data in the beginning columns followed
by a matrix of values by age, length, etc.
}
\author{
Kelli Faye Johnson
}
