% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3sim-package.r
\docType{package}
\name{ss3sim-package}
\alias{ss3sim}
\alias{ss3sim-package}
\title{ss3sim: Fisheries stock assessment simulation testing with Stock Synthesis}
\description{
The \pkg{ss3sim} \R package is designed to
simplify the steps needed to generate beautiful simulation output from the
Stock Synthesis statistical catch-at-age stock assessment framework.
}
\details{
An \pkg{ss3sim} simulation requires three types of input: (1) a base
model of the underlying truth (a Stock Synthesis operating model), (2) a base
model of how you will assess that truth (a Stock Synthesis estimation model),
(3) and a data frame specifying how you want to manipulate the
operating and estimation models.

You can find example an operating and estimation model
within the package data (\verb{inst/extdata/models/}).

To carry out \pkg{ss3sim} simulations with the CRAN version of ss3sim,
you will need to have Stock Synthesis installed on your computer and
the binary needs to be in the path that \R sees.
See
\href{https://nmfs-stock-synthesis.github.io/ss-documentation/Getting_Started_SS.html#the-path-approach}{Stock Synthesis, Getting Started} and
the \href{https://ss3sim.github.io/ss3sim/articles/introduction.html}{Introduction vignette}
for instructions on
making sure Stock Synthesis will work from within \R.

The main \pkg{ss3sim} functions are divided into three types:
\enumerate{
\item \verb{change_*()} and \verb{sample_*()} functions that manipulate Stock Synthesis
configuration files. These manipulations generate an underlying "truth"
(operating models) and control our assessment of those models (estimation
models).
\itemize{
\item \code{\link[=change_f]{change_f()}}: Controls fishing mortality.
}

\item \code{\link[=change_tv]{change_tv()}}: Adds time-varying features. For
example, time-varying natural mortality, growth, or selectivity.

\item \code{\link[=sample_lcomp]{sample_lcomp()}}: Controls how length-composition
data are sampled.

\item \code{\link[=sample_agecomp]{sample_agecomp()}}: Controls how age-composition
data are sampled.

\item \code{\link[=sample_index]{sample_index()}}: Controls how the fishery and
survey indices are sampled.

\item \code{\link[=change_e]{change_e()}}: Controls which and how parameters are
estimated.

\item \code{\link[=change_retro]{change_retro()}}: Controls the number of years to
discard for a retrospective analysis.

\item \code{\link[=change_rec_devs]{change_rec_devs()}}: Substitutes recruitment
deviations.

\item \code{\link[=change_lcomp_constant]{change_lcomp_constant()}}: Set the robustification constant
for length composition data.

\item \code{\link[=change_tail_compression]{change_tail_compression()}}: Replace tail compression value
for length composition data.
}
\enumerate{
\item \verb{run_*()} functions that conduct simulations. These functions
generate a folder structure, call manipulation functions, run Stock Synthesis
as needed, and save the output.
\itemize{
\item \code{\link[=run_ss3sim]{run_ss3sim()}}: Main function to run \pkg{ss3sim}
simulations.
}

\item \code{\link[=ss3sim_base]{ss3sim_base()}}: Underlying base simulation
function. Can also be called directly.
}
\enumerate{
\item \verb{get_*()} functions for synthesizing the output.
\itemize{
\item \code{\link[=get_results_scenario]{get_results_scenario()}}: Extract the results for a
single scenario.
}

\item \code{\link[=get_results_all]{get_results_all()}}: Extract results from a series
of scenarios.
}

See the introductory vignette \code{vignette("introduction", package = "ss3sim")}
for more extensive explanation of how to use the \pkg{ss3sim} \R package.

\pkg{ss3sim} was developed by graduate students and post doctoral researchers
at the University of Washington (School of Aquatic and Fishery Sciences and
Quantitative Ecology and Resource Management departments) and Simon Fraser
University. The authors of individual functions are listed within the
function documentation and all contributors are listed in the
\code{DESCRIPTION} file.

If you use \pkg{ss3sim} in a publication, please cite the package as
indicated by running \code{citation("ss3sim")} in the \R console.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ss3sim/ss3sim}
  \item Report bugs at \url{https://github.com/ss3sim/ss3sim/issues}
}

}
\author{
\strong{Maintainer}: Kelli F. Johnson \email{kelli.johnson@noaa.gov} (\href{https://orcid.org/0000-0002-5149-451X}{ORCID})

Authors:
\itemize{
  \item Sean C. Anderson (\href{https://orcid.org/0000-0001-9563-1937}{ORCID})
  \item Kathryn L. Doering (\href{https://orcid.org/0000-0002-0396-7044}{ORCID})
  \item Cole C. Monnahan (\href{https://orcid.org/0000-0003-0871-6700}{ORCID})
  \item Christine C. Stawitz (\href{https://orcid.org/0000-0003-3122-4501}{ORCID})
}

Other contributors:
\itemize{
  \item Curry Cunningham [contributor]
  \item Allan Hicks [contributor]
  \item Felipe Hurtado-Ferro [contributor]
  \item Peter Kuriyama [contributor]
  \item Roberto Licandeo [contributor]
  \item Carey McGilliard [contributor]
  \item Giancarlo H. Moron Correa [contributor]
  \item Melissa Murdian [contributor]
  \item Kotaro Ono [contributor]
  \item Merrill Rudd [contributor]
  \item Cody Szuwalski [contributor]
  \item Ian G. Taylor [contributor]
  \item Juan Valero [contributor]
  \item Athol Whitten [contributor]
  \item Kiva L. Oken [contributor]
}

}
\keyword{internal}
