% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss3sim.r
\docType{package}
\name{ss3sim}
\alias{ss3sim}
\title{ss3sim: Fisheries stock assessment simulation testing with Stock Synthesis}
\description{
The \pkg{ss3sim} \R package is designed to facilitate rapid, reproducible,
and flexible simulation with the widely-used Stock Synthesis (SS)
statistical catch-at-age stock assessment framework.
}
\details{
An \pkg{ss3sim} simulation requires three types of input: (1) a base
model of the underlying truth (an SS operating model), (2) a base
model of how you will assess that truth (an SS estimation model),
(3) and a data frame specifying how you want to manipulate the
operating and estimation models.

You can find example an operating and estimation model
within the package data (\code{inst/extdata/models/}).

To carry out \pkg{ss3sim} simulations
with the CRAN version of ss3sim, you will need to have SS installed on
your computer and the binary needs to be in the path that \R sees. See the
section "Installing the ss3sim R package" in the vignette
\code{vignette("ss3sim-vignette")} for instructions on installing SS. See
the Appendix A "Putting SS in your path" in the vignette for instructions on
making sure SS will work from within \R.

The main \pkg{ss3sim} functions are divided into three types:

1. \code{change} and \code{sample} functions that manipulate SS
configuration files. These manipulations generate an underlying "truth"
(operating models) and control our assessment of those models (estimation
models).
\itemize{
\item \code{\link{change_f}}: Controls fishing mortality.

\item \code{\link{change_tv}}: Adds time-varying features. For
example, time-varying natural mortality, growth, or selectivity.

\item \code{\link{sample_lcomp}}: Controls how length-composition
data are sampled.

\item \code{\link{sample_agecomp}}: Controls how age-composition
data are sampled.

\item \code{\link{sample_index}}: Controls how the fishery and
survey indices are sampled.

\item \code{\link{change_e}}: Controls which and how parameters are
estimated.

\item \code{\link{change_retro}}: Controls the number of years to
discard for a retrospective analysis.

\item \code{\link{change_rec_devs}}: Substitutes recruitment
deviations.

\item \code{\link{change_lcomp_constant}}: Set the robustification constant
for length composition data.

\item \code{\link{change_tail_compression}}: Replace tail compression value
for length composition data.
}

2. \code{run} functions that conduct simulations. These functions
generate a folder structure, call manipulation functions, run SS
as needed, and save the output.
\itemize{
\item \code{\link{run_ss3sim}}: Main function to run \pkg{ss3sim}
simulations.

\item \code{\link{ss3sim_base}}: Underlying base simulation
function. Can also be called directly.
}

3. \code{get} functions for synthesizing the output.
\itemize{
\item \code{\link{get_results_scenario}}: Extract the results for a
single scenario.

\item \code{\link{get_results_all}}: Extract results from a series
of scenarios.
}

See the introductory vignette \code{vignette("introduction",
package = "ss3sim")} for
more extensive explanation of how to use the \pkg{ss3sim} \R package.

\pkg{ss3sim} was developed by graduate students and post doctoral researchers
at the University of Washington (School of Aquatic and Fishery Sciences and
Quantitative Ecology and Resource Management departments) and Simon Fraser
University. The authors of individual functions are listed within the
function documentation and all contributors are listed in the
\code{DESCRIPTION} file.

If you use \pkg{ss3sim} in a publication, please cite the package as
indicated by running \code{citation("ss3sim")} in the \R console.
}
