% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SS_varadjust2.R
\name{SS_varadjust2}
\alias{SS_varadjust2}
\title{Modify variance and sample size adjustments in the control file}
\usage{
SS_varadjust2(dir = "C:/myfiles/mymodels/myrun/",
  ctlfile = "control.ss_new", newctlfile = "control_modified.ss",
  keyword = "Variance_adjustments", newtable = NULL, newrow = NULL,
  rownumber = NULL, maxcols = 100, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{dir}{Directory with control file to change.}

\item{ctlfile}{Control file name. Default="control.ss_new".}

\item{newctlfile}{Name of new control file to be written.
Default="control_modified.ss".}

\item{keyword}{Keyword to use as reference for start of section on
variance adjustments}

\item{newtable}{Optional table of new variance adjustment values}

\item{newrow}{Optional vector of new variance adjustment values for a particular row}

\item{rownumber}{Which of the 6 rows to replace with 'newrow' if present?}

\item{maxcols}{Maximum number of columns to search among (may need to
increase from default if you have a huge number of fleets)}

\item{overwrite}{Overwrite file if it exists?}

\item{verbose}{TRUE/FALSE switch for amount of detail produced by function.
Default=TRUE.}
}
\description{
Function was adapted from \code{\link{SS_varadjust}} as the function 
was not fully developed at the time it was needed for this package
}
\seealso{
\code{\link{SS_parlines}}, \code{\link{SS_changepars}}
}
\author{
Ian G. Taylor, Gwladys Lambert
}
