% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vbgf_func}
\alias{vbgf_func}
\title{Predict length given VBGF parameters}
\usage{
vbgf_func(L1, L2, k, ages, a3)
}
\arguments{
  \item{L1}{description}

  \item{L2}{description}

  \item{k}{description}

  \item{ages}{description}

  \item{a3}{description}
}
\description{
External estimation procedure for von Bertalanffy growth
}

