% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_input.r
\name{verify_input}
\alias{verify_input}
\title{Verify and standardize Stock Synthesis input files}
\usage{
verify_input(model_dir, type = c("om", "em"))
}
\arguments{
\item{model_dir}{Directory name for model. This folder should contain the
\code{.ctl}, \code{.dat}, files etc.}

\item{type}{One of "om" or "em" for operating or estimating model.}
}
\value{
Nothing is returned from this function. Instead,
file are changed and saved to the disk.
}
\description{
Verify the contents of
operating model (\code{OM}) and
estimation model (\code{EM}) folders, i.e.,
check that the necessary SS input files are available.
If the contents are correct,
the \code{.ctl} and \code{.dat} files are renamed to standardized names and
the \code{starter.ss} file is updated to reflect these names.
If the contents are incorrect,
then a warning is issued and the simulation is aborted.
}
\examples{
# Create a temporary folder for the output:
temp_path <- file.path(tempdir(), "ss3sim-verify-example")
dir.create(temp_path, showWarnings = FALSE)

d <- system.file("extdata", "models", package = "ss3sim")

om <- file.path(d, "cod-om")
em <- file.path(d, "cod-em")

file.copy(om, temp_path, recursive = TRUE)
file.copy(em, temp_path, recursive = TRUE)

# Verify the correct files exist and change file names:
verify_input(model_dir = file.path(temp_path, "cod-om"), type = "om")
verify_input(model_dir = file.path(temp_path, "cod-em"), type = "em")
unlink(temp_path, recursive = TRUE)
}
\author{
Curry James Cunningham; modified by Sean C. Anderson
}
