% Generated by roxygen2 (4.0.1): do not edit by hand
\name{verify_plot_arguments}
\alias{verify_plot_arguments}
\title{A helper function to check the correct input for the plotting functions.}
\usage{
verify_plot_arguments(data, x, y, horiz, horiz2, vert, vert2, color,
  relative.error, axes.free)
}
\description{
A helper function to check the correct input for the plotting functions.
}
\author{
Cole Monnahan
}
