% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_comps.R
\name{weight_comps}
\alias{weight_comps}
\title{Weight the composition data in the EM}
\usage{
weight_comps(
  method = c("MI", "Francis", "DM"),
  init_run = TRUE,
  main_run = TRUE,
  niters_weighting = 1,
  iter,
  scen,
  bias_adjust = FALSE,
  hess_always = FALSE,
  fleets,
  ...
)
}
\arguments{
\item{method}{Options for the weighting method are
"MI" (for MacAllister-Ianelli), "Francis", or "DM" (for dirichlet
multinomial). For now, only 1 option can be used.}

\item{init_run}{A logical value specifying if the model should be run before
iterative reweighting. Only relavent to \code{method = c("MI", "Francis")}
because they require initial estimates that are compared to a theoretical
distribution external to SS.}

\item{main_run}{A logical value specifying if the model should be run to
weight the composition data after the appropriate weightings are found.
If \code{FALSE}, the final values found for the weightings will only
be put into the control file and not used to generate parameter estimates.}

\item{niters_weighting}{The number of times you want to tune the model.}

\item{iter}{The iteration. A numeric value. This function assumes model
files are in a directory relative to the path \code{"scen/iter/EM"} as in
ss3sim.}

\item{scen}{The scenario, named as in ss3sim. This function assumes model
files are in a directory relative to the path \code{"scen/iter/EM"} as in
ss3sim.}

\item{bias_adjust}{Run bias adjustment first?}

\item{hess_always}{If \code{TRUE}, the Hessian will always be calculated.
If \code{FALSE}, the Hessian will only be calculated for
bias-adjustment runs thereby saving time.}

\item{fleets}{A vector of numeric values specifying which fleets you want to
tune.}

\item{...}{Anything else to pass to \code{\link{run_ss3model}}.}
}
\description{
Weight age- or length-composition data in the estimation model (EM) using a
tuning method (MacAllister-Ianelli or Francis) or a distribution
(dirichlet multinomial).
}
