\name{add_time_varying_features}
\alias{add_time_varying_features}
\title{Methods to include time varying param features
Documentation not complete
Include_timevarying_param accept three different options
1. using the "block" option for time varying param
2. using the "dev" option for time varying param
3. using the "env" option for time varying param}
\usage{
  add_time_varying_features(working_folder = working_folder,
    ctl.file.in = "control.ss_new",
    ctl.file.out = "YTF.ctl", Dat.file = "YTF.dat",
    par.file = "ss3.par", Nblocks, Blockpattern, Dev,
    how.time.varying = "env")
}
\description{
  Methods to include time varying param features
  Documentation not complete Include_timevarying_param
  accept three different options 1. using the "block"
  option for time varying param 2. using the "dev" option
  for time varying param 3. using the "env" option for time
  varying param
}

