% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bias.R
\name{calculate_bias}
\alias{calculate_bias}
\title{Calculate bias adjustment for recruitment deviations}
\usage{
calculate_bias(dir, ctl_file_in)
}
\arguments{
\item{dir}{A character string specifying the path to the folder with the
results from a stock assessment model run. Paths are passed without a
terminal slash.}

\item{ctl_file_in}{A string providing the path to the input SS \code{.ctl} file.}
}
\value{
A list of bias adjustment parameters.
}
\description{
Bias adjustment is performed to ensure that the only the most informative
data available are used when estimating recruitment deviations.
This process involves running the estimation method with the hessian
both before this function and then running the estimation method again with
the new values. Estimation files in the original folder will be deleted to
ensure that convergence afterwards is not based on the input run.
These files are permanently archived in the \verb{bias_[0-9]\{2\}} folder
within the directory. Only the three middle steps listed below are performed
using \code{calculate_bias}, and the rest of the steps must be performed
externally.
}
\details{
\itemize{
\item Estimate recruitment and the standard error about those estimates.
\item Correct the estimates given their estimated uncertainty using a ramp.
\item Save a new control file.
\item Move original estimation files.
\item Estimate model parameters.
}
}
\author{
Kelli Faye Johnson
}
