% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case-write.R
\name{case_comp}
\alias{case_comp}
\title{Write a case file for length- or age-composition data}
\usage{
case_comp(fleets = 1, Nsamp = NULL, years = NULL, cpar = 2, type, case,
  spp)
}
\arguments{
\item{fleets}{Vector of fleet numbers, where the order of
\code{fleets} will dictate the order of all remaining arguments.}

\item{Nsamp}{A list containing vectors of sample sizes for each
year for each fleet.}

\item{years}{A list containing vectors of years for each year for each fleet.}

\item{cpar}{A vector of cpar for each fleet.}

\item{type}{A character value of \code{"agecomp"} or \code{"lcomp"},
to write age- or length-composition specifications, respectively.
Argument can be a vector (e.g., \code{c("agecomp", "lcomp")}) if you want
the case files to be the same for length and age compositions.}

\item{case}{The case you want to write to. E.g. \code{"M"}.}

\item{spp}{A vector of character values argument specifying the species.}
}
\description{
Write a case file for length- or age-composition data
}
\examples{
case_comp(fleets = 1:2, case = 30, spp = "cod",
  Nsamp = list(rep(10, 40), rep(10, 25)),
  years = list(61:100, 76:100), cpar = 2:1, type = "agecomp")
done <- file.remove("agecomp30-cod.txt")
}

