% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case-write.R
\name{case_index}
\alias{case_index}
\title{Write a case file for index data to the disk.}
\usage{
case_index(fleets = 1, years = NULL, sd = 2, case, spp)
}
\arguments{
\item{fleets}{Vector of fleet numbers, where the order of
\code{fleets} will dictate the order of all remaining arguments.}

\item{years}{A list containing vectors of years for each year for each fleet.}

\item{sd}{A vector of standard deviations for each fleet.}

\item{case}{The case you want to write to. E.g. \code{'I'}.}

\item{spp}{A vector of character values argument specifying the species.}
}
\description{
Write a case file for index data to the disk.
}
\examples{
case_index(fleets = 2, case = 1, spp = "cod", years = list(7:10), sd = 0.1)
done <- file.remove("index1-cod.txt")
}

