% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_agecomp.R
\name{change_agecomp}
\alias{change_agecomp}
\title{(Depreciated) Sample age compositions from expected values}
\usage{
change_agecomp(...)
}
\arguments{
\item{...}{Arguments that get passed to \code{\link{sample_agecomp}}.}
}
\description{
\code{change_agecomp} is a depreciated function. Please use
\code{\link{sample_agecomp}} instead. \code{change_agecomp} will be removed
in the next major version.
}

