% Generated by roxygen2 (4.0.2): do not edit by hand
\name{change_bin}
\alias{change_bin}
\title{Change the data that is available as output from an SS operating model.}
\usage{
change_bin(file_in, file_out, type = "len", bin_vector = NULL,
  pop_bin = NULL, fleet_dat = NULL, write_file = TRUE)
}
\arguments{
  \item{file_in}{A character value giving the location of
  an SS \code{.dat} file to input.}

  \item{file_out}{A character value giving the location of
  an SS \code{.dat} file to output.}

  \item{type}{A vector that can take the one or all of the
  following entries: \code{"len"}, \code{"age"},
  \code{"cal"}, \code{"mla"}, \code{"mwa"}. \code{type}
  controls what data structures the function acts on, with
  \code{"len"} augmenting the length composition data,
  \code{"age"} augmenting the age composition, \code{"cal"}
  augmenting the conditional age at length, \code{"mla"}
  augmenting the mean length at age data, and \code{"mwa"}
  augmenting the mean weight at age data. Default is
  \code{"len"}.}

  \item{bin_vector}{A named list of vectors. Named elements
  provide the desired bin structure for each data
  \code{type}. If \code{type = "len"} or \code{type =
  "age"} \code{bin_vector} should be a vector of new bins.
  For other types in \code{bin_vector} the bins are
  determined from the current or specified length and age
  bins. If names are forgotten, list elements will be named
  according to the order of entries in \code{type}.}

  \item{pop_bin}{A positive real number for the population
  bin width. This option only works if \code{"lbin_method"}
  is set to \code{"2"} in SS3. Default is \code{pop_bin =
  NULL} which leaves the original value.}

  \item{fleet_dat}{A named list of vectors giving the
  desired fleets and years for each data type. The names of
  the list elements should correspond to the \code{type}
  and each present element should have \code{$years} and
  \code{$fleets} giving the vector of desired years and
  fleets. Named elements only need to be specified for
  those required by \code{type}.}

  \item{write_file}{Should the \code{.dat} file be written?
  The new \code{.dat} file will always be returned
  invisibly by the function. Setting \code{write_file =
  FALSE} can be useful for testing. Note that you must
  supply a value to the argument \code{file_out}, but this
  argument can be set to any arbitrary value (such as
  \code{NULL}) if \code{write_file = FALSE}.}
}
\description{
\code{change_bin} alters the bin structure for the age or
length composition data in an SS operating model. Original
data is removed and dummy data is added at the appropriate
bin sizes to the SS \code{.dat} file, causing SS to record
age or length composition data in the appropriate bins when
the operating model is run. Additionally the code will
introduce dummy conditional length-at-age or size- (weight
or length) at-age data to the \code{.dat} file. For each
data type altered, \code{change_bin} will add data in a
full factorial manner, working with existing fleets for all
years; potentially adding many rows of data. Currently,
\code{.dat} files with multiple genders cannot be
manipulated with \code{change_bin}. Use
\code{\link{sample_lcomp}}, \code{\link{sample_agecomp}},
and \code{\link{sample_calcomp}} to reduce the data.
}
\details{
Within the \code{.dat} file, the conditional age-at-length
data is stored in the same matrix as the age composition
data. Thus, it is necessary that the conditional
age-at-length data use the same binning structure as the
age composition data. If \code{type = "caa"} and not
\code{type = c("age", "caa")} the function will add
conditional age-at-length using the binning structure of
the current \code{.dat} file.  Also note that if \code{type
= c("mla", "mwa")} no entries are currently necessary in
the \code{bin_vector}.
}
\examples{
d <- system.file("extdata", package = "ss3sim")
f_in <- paste0(d, "/example-om/data.ss_new")
out <- change_bin(f_in, file_out = NULL,
  type = c("len", "age", "cal", "mla", "mwa"),
  fleet_dat = list(
    "len" = list(years = list(1:3, 1:3), fleets = 1:2),
    "age" = list(years = list(1:3, 1:3), fleets = 1:2),
    "cal" = list(years = list(1:3, 1:3), fleets = 1:2),
    "mla" = list(years = list(1:3, 1:3), fleets = 1:2),
    "mwa" = list(years = list(1:3, 1:3), fleets = 1:2)),
  bin_vector = list(
    "len" = seq(2, 8, 2),
    "age" = seq(2, 8, 2)),
  write_file = FALSE)
print(out$agebin_vector)
print(tail(out$agecomp))
print(head(out$MeanSize_at_Age_obs))
}
\author{
Ian Taylor, Sean Anderson, Kelli Johnson
}
\seealso{
\code{\link{sample_lcomp}}, \code{\link{sample_agecomp}}
}

