% Generated by roxygen2 (4.0.2): do not edit by hand
\name{change_bin}
\alias{change_bin}
\title{Change the data that is available as output from an SS operating model.}
\usage{
change_bin(file_in, file_out, bin_vector, type = "len", pop_bin = NULL,
  write_file = TRUE)
}
\arguments{
  \item{file_in}{A character value giving the location of
  an SS \code{.dat} file to input.}

  \item{file_out}{A character value giving the location of
  an SS \code{.dat} file to output.}

  \item{bin_vector}{A named list of vectors. Named elements
  provide the desired bin structure for each data
  \code{type}, where the \code{type} is an additional
  argument. List should be named to map the appropriate
  vector to each \code{type}. If names are forgotten, list
  elements will be named according to the order of entries
  in \code{type}. Each vector of bins is substituted into
  the \code{.dat} file.}

  \item{type}{A vector that can take the one or all of the
  following entries: \code{"len"}, \code{"age"},
  \code{"cal"}, \code{"mla"}, \code{"mwa"}. \code{type}
  controls what data structures the function acts on, with
  \code{"len"} augmenting the length composition data,
  \code{"age"} augmenting the age composition, \code{"cal"}
  augmenting the conditional age at length, \code{"mla"}
  augmenting the mean length at age data, and \code{"mwa"}
  augmenting the mean weight at age data. Default is
  \code{"len"}.}

  \item{pop_bin}{Choose a real number to choose the
  population bin width. This option only works if
  \code{"lbin_method"} is set to \code{"2"}. Default is
  \code{"NULL"} which leaves the original value.}

  \item{write_file}{Should the \code{.dat} file be written?
  The new \code{.dat} file will always be returned
  invisibly by the function. Setting \code{write_file =
  FALSE} can be useful for testing. Note that you must
  supply a value to the argument \code{file_out}, but this
  argument can be set to any arbitrary value (such as
  \code{NULL}) if \code{write_file = FALSE}.}
}
\description{
\code{change_bin} alters the bin structure for the age or
length composition data in an SS operating model. Original
data is removed and dummy data is added at the appropriate
bin sizes to the SS \code{.dat} file, causing SS to record
age or length composition data in the appropriate bins when
the operating model is run. Additionally the code will
introduce dummy conditional length-at-age or size- (weight
or length) at-age data to the \code{.dat} file. For each
data type altered, \code{change_bin} will add data in a
full factorial manner, working with existing fleets,
seasons, population ages, and ageing error matrices;
potentially adding many rows of data. Currently,
\code{.dat} files with multiple genders cannot be
manipulated with \code{change_bin}. Use
\code{\link{sample_lcomp}}, \code{\link{sample_agecomp}},
and \code{\link{sample_ccomp}} to reduce the data.
}
\details{
Within the \code{.dat} file, the conditional age-at-length
data is stored in the same matrix as the age composition
data. Thus, it is necessary that the conditional
age-at-length data use the same binning structure as the
age composition data. If \code{type = "caa"} and not
\code{type = c("age", "caa")} one will only add conditional
age-at-length using the binning structure of the current
\code{.dat} file.  \code{change_bin} will only add
conditional age-at-length with a binning width equal to 1
age bin (i.e. Lbin_lo == Lbin_hi).  Also note that if
\code{type = c("mla", "mwa")} no entries are currently
necessary in the \code{bin_vector}.
}
\author{
Ian Taylor, Sean Anderson, Kelli Johnson
}
\seealso{
\code{\link{sample_lcomp}}, \code{\link{sample_agecomp}}
}

