% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_catch.R
\name{change_catch}
\alias{change_catch}
\title{Change catch in the Stock Synthesis data list}
\usage{
change_catch(dat_list, ctl_list)
}
\arguments{
\item{dat_list}{A data file read in using \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}}.}

\item{ctl_list}{A control file read in using \code{\link[r4ss:SS_readctl]{r4ss::SS_readctl()}}.
The start and end year of the resulting data list will be based on years
with positive fishing mortality values, and
equilibrium catches will be non-zero only if
there is a equilibrium fishing mortality parameter for that fleet and month
combination.}
}
\value{
A modified Stock Synthesis data file as a list in R.
}
\description{
Change catch in the data so at least all combinations of
fleet, month, and year,
needed for catch are available.
Equilibrium years are generated if
there are equilibrium parameters in the control list.
}
\seealso{
\code{\link[=change_f]{change_f()}} changes the fishing mortality, \eqn{F}, parameters
using the control file, but these \eqn{F} values will only be implemented
for years with corresponding entries in the Stock Synthesis data file.
Thus, this function must be implemented after \code{\link[=change_f]{change_f()}}.
}
\author{
Kathryn L. Doering
}
