% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_e.r
\name{change_e}
\alias{change_e}
\title{Methods to alter the parameters estimated in a Stock Synthesis model}
\usage{
change_e(
  ctl_file_in = "em.ctl",
  ctl_file_out = "em.ctl",
  dat_list = NULL,
  for_file_in = "forecasts.ss",
  par_name = NULL,
  par_int = "NA",
  par_phase = "NA",
  forecast_num = 0,
  verbose = FALSE,
  natM_type = NULL,
  natM_n_breakpoints = NULL,
  natM_lorenzen = NULL,
  natM_val = NULL
)
}
\arguments{
\item{ctl_file_in}{A string providing the path to the input SS \code{.ctl} file.}

\item{ctl_file_out}{A string providing the path to the output SS control file. If the value is \code{NULL}, the file will not be written to the disk.}

\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{for_file_in}{A string providing the path to the input SS
\code{forecast.ss} file.}

\item{par_name}{A vector of character values corresponding to parameter
names that you wish to initialize at different values or change the phase
in which they are estimated. Entries are searched for in
\code{ctl_file_in}, and therefore, it is best to use full parameter
names as they are specified in the file.}

\item{par_int}{A vector of initial values, one for each entry in
\code{par_name}. Values of \code{NA} leave the INIT value for
that parameter at the value found in the \code{.ctl} file.}

\item{par_phase}{A vector of phase values, one for each parameter in
\code{par_name}.  Values can be \code{NA} if you do not wish to change
the phase for a given parameter. Negative values will fix the parameter
at the INIT value.}

\item{forecast_num}{Number of years to perform forecasts. For those years,
the data will be removed from the \code{dat_list}, enabling Stock Synthesis to
generate forecasts rather than use the data to fit the model.}

\item{verbose}{When \code{TRUE} messages will be returned
from the function. Often useful for debugging. The default
is \code{FALSE}.}

\item{natM_type}{Deprecated. Should have value NULL.}

\item{natM_n_breakpoints}{Deprecated. Should have value NULL.}

\item{natM_lorenzen}{Deprecated. Should have value NULL.}

\item{natM_val}{Deprecated. Should have value NULL.}
}
\value{
Altered versions of Stock Synthesis \code{.ctl} and \code{forecast.ss} files are written
to the disk and the altered \code{dat_list} is returned invisibly.
}
\description{
Takes Stock Synthesis
\code{.ctl} and \code{forecast.ss} files, along with
a list structure which houses the data file as read in by
\code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}}
and changes which parameters are estimated, how natural mortality is
estimated, and if forecasts are performed. The function can be called by
itself or within \code{\link[=run_ss3sim]{run_ss3sim()}} to alter an estimation model
\code{.ctl} file.
}
\examples{
d <- system.file("extdata", "models", "cod-om", package = "ss3sim")
change_e(
  ctl_file_in = file.path(d, "codOM.ctl"),
  ctl_file_out = file.path(tempdir(), "change_e.ctl"),
  dat_list = codomdat,
  for_file_in = file.path(d, "forecast.ss"),
  par_name = c("_steep", "Size_DblN_peak_Fishery(1)"),
  par_int = c(0.3, 40), par_phase = c(3, 2),
  forecast_num = 0
)
# clean up the temporary files
file.remove(file.path(tempdir(), "change_e.ctl"))
}
\seealso{
Other change functions: 
\code{\link{change_data}()},
\code{\link{change_em_binning}()},
\code{\link{change_f}()},
\code{\link{change_o}()},
\code{\link{change_retro}()},
\code{\link{change_tv}()}
}
\author{
Kelli F. Johnson
}
\concept{change functions}
