% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_em_ageerr.R
\name{change_em_ageerr}
\alias{change_em_ageerr}
\title{Change aging error matrix in estimation model age composition}
\usage{
change_em_ageerr(dat_list, outfile = NULL, fleets, definition)
}
\arguments{
\item{dat_list}{A Stock Synthesis  data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the
file takes time and computing resources.}

\item{fleets}{A vector of integers specifying the fleets to include. The
order of the fleets corresponds to the order of the \code{definition} argument.
An entry of \code{fleets=NULL} leads to no ageing error definitions being
changed in the EM.}

\item{definition}{A vector of integers, of the same length as \code{fleets},
specifying the aging error definitions to include. The order of the
definitions corresponds to the order of the \code{fleets} argument. The aging
error definitions must already input in the operating model (OM) data file.}
}
\description{
\code{change_em_ageerr()} alters the fleet specific aging error definition in
a stock synthesis estimation model (EM). The original aging error definition
in the age composition observation section of the data file is changed
according to the user's specification.
}
\details{
The arguments \code{fleets} and \code{definition} are both vectors, or a list of vectors
that must be of equal length. The order in which the fleets are specified in
the \code{fleets} vector should correspond to the intended new aging error
definition in the \code{definition} vector. Currently this function only allows
for fleet specific changes to  aging error observations, not year specific.

Users need to input all aging error matrices being considered into the
operating model (OM) data file, as this function changes an aging error
definition in the age composition observation section of the dat file
and does not allow a new aging error matrix to be added.

This function can also be called in \code{\link[=run_ss3sim]{run_ss3sim()}} by inputting a vector of
aging error definitions, corresponding to the aging error definition in each
scenario, into a column labeled \code{ca.definition.1} in the \code{simdf} data frame.
Where the number in \code{ca.definition.#} corresponds to the fleet. e.g.-
\code{ca.definition.1 <- c(3, 2)} will change the aging error definition, in the EM,
of fleet 1 to 3 in the first scenario and 2 in the second scenario.

Aging error definitions can also be changed in the OM in \code{\link[=run_ss3sim]{run_ss3sim()}}
by inputting a vector of aging error definitions, corresponding to the
aging error definition in each scenario, into a column labeled
\code{sa.Ageerr.1} in the \code{simdf} data frame. Where the number in \code{sa.Ageerr.#}
corresponds to the fleet. e.g.- \code{sa.Ageerr.1 <- c(3, 2)} will change the
aging error definition, in the OM, of fleet 1 to 3 in the first scenario
and 2 in the second scenario.
}
\examples{
d <- system.file("extdata", package = "ss3sim")
f_in <- file.path(d, "models", "cod-om", "codOM.dat")
dat_list <- r4ss::SS_readdat(f_in, verbose = FALSE)

# Adding in second aging error definition to the data file
dat_list$N_ageerror_definitions <- 2
dat_list$ageerror <- rbind(dat_list$ageerror,dat_list$ageerror)

# An example changing the aging error observation for multiple fleets
new_dat <- change_em_ageerr(
  dat_list = dat_list, outfile = NULL, fleets = c(1, 2), definition = c(2, 2)
)

# An example changing the aging error observations for a single fleet
new_dat2 <- change_em_ageerr(
  dat_list = dat_list, outfile = NULL, fleets = 2, definition = 2
)

rm(d, f_in, dat_list, new_dat, new_dat2)
}
\seealso{
Other change functions: 
\code{\link{change_data}()},
\code{\link{change_em_binning}()},
\code{\link{change_e}()},
\code{\link{change_f}()},
\code{\link{change_o}()},
\code{\link{change_retro}()},
\code{\link{change_tv}()}
}
\author{
Derek W. Chamberlin
}
\concept{change functions}
