% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/change_em_binning.r
\name{change_em_binning}
\alias{change_em_binning}
\title{Change population and observed length composition bins in an SS estimation
model}
\usage{
change_em_binning(file_in, file_out, lbin_method = NULL, bin_vector,
  write_file = TRUE)
}
\arguments{
  \item{file_in}{A character value giving the location of
  an SS \code{ss3.dat} file to input.}

  \item{file_out}{A character value giving the location of
  an SS \code{ss3.dat} file to output.}

  \item{lbin_method}{A numeric value of either
  \code{"NULL","1","2","3"} to change the lbin_method for
  the population bin. Only supports either
  \code{"NULL","1","2"} at the moment. NULL means to keep
  it unchanged}

  \item{bin_vector}{A numeric vector of bins to substitute
  into the \code{ss3.dat} file.}

  \item{write_file}{Should the \code{.dat} file be written?
  The new \code{.dat} file will always be returned
  invisibly by the function. Setting \code{write_file =
  FALSE} can be useful for testing. Note that you must
  supply a value to the argument \code{file_out}, but this
  argument can be set to any arbitrary value (such as
  \code{NULL}) if \code{write_file = FALSE}.}
}
\description{
\code{change_em_binning} alters the bin structure for the
population and length composition data in an SS estimation
model. It is done by taking the original length composition
info from the EM \code{ss3.dat} then changing according to
the user's specification.
}
\examples{
d <- system.file("extdata", package = "ss3sim")
f_in <- paste0(d, "/example-om/data.ss_new")
l <- change_em_binning(f_in, file_out = NULL, lbin_method=1,
  bin_vector = seq(8,30,by=1), write_file = FALSE)
print(l$lbin_vector)
print(head(l$lencomp))

l <- change_em_binning(f_in, file_out = NULL, lbin_method=1,
  bin_vector = seq(10,27,by=2), write_file = FALSE)
print(l$lbin_vector)
print(head(l$lencomp))
}
\author{
Kotaro Ono
}
\seealso{
\code{\link{sample_lcomp}}, \code{\link{sample_agecomp}}

Other change functions: \code{\link{change_data}};
\code{\link{change_f}}; \code{\link{change_maturity}};
\code{\link{change_retro}}; \code{\link{change_tv}}
}

