% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/change_f.r
\name{change_f}
\alias{change_f}
\title{Alter the fishing mortality (F) values in an \code{ss3.par} file}
\usage{
change_f(years, years_alter, fvals, file_in = "ss3.par",
  file_out = "ss3.par")
}
\arguments{
\item{years}{*Vector of years for which F values are specified,
if there is more than one fleet or season the catches must be ordered by
season:year:fishey (e.g., season1year1fishery1, season2year1fishery1,
season1year2fishery1). The actual vector does not have to correspond to
true years but must be the correct length (e.g., instead of
\code{2000:2004} you can use \code{1:5}). Use this argument to create
an index to old values. \code{years_alter} will use values in this vector.
For example, with two seasons and one fishery that operates for 4 years
you could use the following: \code{1:8}.}

\item{years_alter}{*Vector of years for the which F values will be altered.
If there is more than one fishery or season, use the mapping system
created in \code{years} because actual year values cannot be recycled.
For example, to change the second season of the second year in the
example above, use: \code{4}.}

\item{fvals}{*Vector of F values to be entered into \code{ss3.par} file}

\item{file_in}{Input SS3 par file.}

\item{file_out}{Output SS3 par file.}
}
\value{
A modified SS3 \code{.par} file.
}
\description{
Takes an SS3 \code{.par} file and changes the F values for specified years.
If used with \code{\link{run_ss3sim}} the case file should be named
\code{F}. A suggested (default) case letter is \code{F}.
}
\section{Which arguments to specifiy in case files}{

All function argument descriptions that start with an asterisk (*) will be passed
through the case files to \code{\link{run_ss3sim}}. If one of these arguments
is not specified in a case file, then a value of \code{NULL} will be passed,
which may or may not be an appropriate value. Other arguments will be ignored
if specified.
}
\examples{
# Create a temporary folder for the output:
temp_path <- file.path(tempdir(), "ss3sim-f-example")
dir.create(temp_path, showWarnings = FALSE)

# Find the example .par file in the package data:
d <- system.file("extdata", package = "ss3sim")
par_file <- paste0(d, "/change_f/ss3.par")

change_f(years = 1:49, years_alter = 2, fvals = 9999, file_in =
par_file, file_out = paste0(temp_path, "/test.par"))
}
\author{
Curry James Cunningham
}
\seealso{
Other change functions: \code{\link{change_data}};
  \code{\link{change_em_binning}}; \code{\link{change_e}};
  \code{\link{change_maturity}};
  \code{\link{change_retro}}; \code{\link{change_tv}}
}

