% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_index.r
\name{change_index}
\alias{change_index}
\title{(Depreciated) Sample the biomass with observation error}
\usage{
change_index(...)
}
\arguments{
\item{...}{Arguments that get passed to \code{\link{sample_index}}.}
}
\description{
\code{change_index} is a depreciated function. Please use
\code{\link{sample_index}} instead. \code{change_index} will be removed
in the next major version.
}

