% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_lcomp.R
\name{change_lcomp}
\alias{change_lcomp}
\title{(Depreciated) Sample length compositions from expected values}
\usage{
change_lcomp(...)
}
\arguments{
\item{...}{Arguments that get passed to \code{\link{sample_lcomp}}.}
}
\description{
\code{change_lcomp} is a depreciated function. Please use
\code{\link{sample_lcomp}} instead. \code{change_lcomp} will be removed
in the next major version.
}

