% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_lcomp_constant.R
\name{change_lcomp_constant}
\alias{change_lcomp_constant}
\title{Set the robustification constant for length-composition data}
\usage{
change_lcomp_constant(lcomp_constant, dat_list, outfile = NULL)
}
\arguments{
\item{lcomp_constant}{The new value to be used. Must be a numeric value, as
a proportion. For example 0.1 means 10 percent. See the Stock Synthesis manual for
further information. A NULL value indicates no action resulting in using
the current value, and a value of 0 will throw an error since that leads to
an error when zeroes exist in the data. Instead use a very small value like
1e-07.}

\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the
file takes time and computing resources.}
}
\value{
A modified Stock Synthesis \code{.dat} file, and that file returned invisibly
(for testing) as a vector of character lines.
}
\description{
This function replaces the robustification value for length-composition data
in a \code{.dat} file that was read in using \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}}
with those specified in \code{lcomp_constant}.
It then writes a new file with name \code{outfile} into the working directory.
}
\details{
The robustification constant is added to both the observed and
expected proportions of length composition data, before being normalized
internally. It is designed to help stabilize the model, but is unclear how
and when to use it for optimal effect. The same value is used for all
length data.
}
\author{
Cole Monnahan
}
