% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_o.R
\name{change_o}
\alias{change_o}
\title{Methods to include parameters in an SS3operating model}
\usage{
change_o(
  change_o_list,
  ctl_file_in = "control.ss_new",
  ctl_file_out = "om.ctl",
  par_name = NULL,
  par_int = NULL,
  verbose = FALSE
)
}
\arguments{
\item{change_o_list}{A list of named vectors. Names correspond to parameters
in the operating model and the vectors correspond to deviations.
Alternatively, \code{par_name} and \code{par_init} can be passed to this
function.}

\item{ctl_file_in}{A string providing the path to the input SS \code{.ctl} file.}

\item{ctl_file_out}{A string providing the path to the output SS control file. If the value is \code{NULL}, the file will not be written to the disk.}

\item{par_name}{A vector of character values corresponding to parameter
names that you wish to initialize at different values or change the phase
in which they are estimated. Entries are searched for in
\code{ctl_file_in}, and therefore, it is best to use full parameter
names as they are specified in the file.}

\item{par_int}{A vector of initial values, one for each entry in
\code{par_name}. Values of \code{NA} leave the INIT value for
that parameter at the value found in the \code{.ctl} file.}

\item{verbose}{When \code{TRUE} messages will be returned
from the function. Often useful for debugging. The default
is \code{FALSE}.}
}
\value{
The function creates modified versions of the  \code{.ctl} files. The
  function also returns \code{change_o_list} invisibly.
}
\description{
\code{change_o} takes an SS3 \code{.ctl}file
and implements parameter value changes that are NOT time varying.
\code{change_o} is specifically set up to work with an operating model
\code{.ctl} file.
}
\section{Specifying the \code{change_o_list}}{


Parameters initial values will change according to the values passed to
\code{change_o_list}. Each parameter should have a single value specified.

Parameter names must be unique and match the full parameter name in the
\code{.ctl} file.
}

\section{Passing arguments to \code{change_o} through \code{\link{run_ss3sim}}}{

(1) add a column called `co.par_name` to the `simdf` that specifies which parameters
you want to change in the OM, each element of this vector needs to be
wrapped in quotations to be later evaluated, e.g., `'c("SR_BH_steep","SR_sigmaR")'`
represents a single entry;
and (2) add an additional column called `co.par_int` to the `simdf` that specifies
INIT values for each parameter in the previous column, e.g., `"c(0.6, 1.0)"`, if
there is more than one value, the vector needs to be wrapped in quotations to be
evaluated later.
}

\seealso{
Other change functions: 
\code{\link{change_data}()},
\code{\link{change_em_binning}()},
\code{\link{change_e}()},
\code{\link{change_f}()},
\code{\link{change_retro}()},
\code{\link{change_tv}()}
}
\author{
Kathryn Doering
}
\concept{change functions}
