% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_year.R
\name{change_startyear}
\alias{change_startyear}
\title{Change start year of the data file}
\usage{
change_startyear(dat_list, firstyear = NULL)
}
\arguments{
\item{dat_list}{A Stock Synthesis  data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{firstyear}{An integer specifying the year to start fitting the model.
The default is \code{NULL}, which will look up the first year
with non-zero catch or non-zero catch-per-unit-effort data in \code{dat_list}.}
}
\value{
A \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}} list with an augmented start year.
}
\description{
Change start year of the data file
}
\author{
Kelli F. Johnson
}
