% Generated by roxygen2 (4.0.1): do not edit by hand
\name{change_tail_compression}
\alias{change_tail_compression}
\title{Replace tail compression value for length composition data.}
\usage{
change_tail_compression(tail_compression, file_in, file_out)
}
\arguments{
  \item{tail_compression}{*The new tail_compression value
  to be used. Must be a numeric value, as a proportion. For
  example 0.1 means 10 percent. See the SS3 manual for
  further information. A NULL value indicates no action, a
  negative value indicates to SS3 to ignore it (not use
  that feature).}

  \item{file_in}{Input SS3 dat file.}

  \item{file_out}{Output SS3 dat file. Typically the same
  as \code{file_in}.}
}
\value{
A modified SS3 \code{.dat} file, and that file returned
invisibly (for testing) as a vector of character lines.
}
\description{
This function replaces the tail compression value for
length composition data in a \code{dat} file
(\code{file_in}) with those specified in
\code{tail_compression}. It then writes a new file with
name \code{file_out} into the working directory.
}
\section{Which arguments to specifiy in case files}{
  All function argument descriptions that start with an
  asterisk (*) will be passed through the case files to
  \code{\link{run_ss3sim}}. If one of these arguments is
  not specified in a case file, then a value of \code{NULL}
  will be passed, which may or may not be an appropriate
  value. Other arguments will be ignored if specified.
}
\examples{
## Create a temporary folder for the output:
temp_path <- file.path(tempdir(), "ss3sim-tail-example")
dir.create(temp_path, showWarnings = FALSE)
## Run the function with built-in data file.
dat_file <- system.file("extdata", "example-om", "data.ss_new",
  package = "ss3sim")
test <- change_tail_compression(tail_compression = .1234, file_in = dat_file,
  file_out = paste0(temp_path, "/test.dat"))
## Look at the changes
test[grep("_tail_compression", test)[1]]
## Clean up the temp files
unlink(temp_path)
}
\author{
Cole Monnahan
}

