% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_year.R
\name{change_year}
\alias{change_year}
\title{Change the years estimated}
\usage{
change_year(dat_list, ctl_list)
}
\arguments{
\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{ctl_list}{A control file read in by \code{\link[r4ss]{SS_readctl}}.}
}
\description{
Keep all of the data in the model but change the years that are estimated
in the model. First year of the model will be first year of non-zero catch.
Main recruitment period starts 1/2 generation time before first year 
of compositional data included in the model.
Late recruitment is the last year of the model by default and cannot be
modified using this function, neither can early recruitment, which starts
in year 1.
}
