% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_data.r
\name{check_data}
\alias{check_data}
\title{Check that the Stock Synthesis data file looks correct}
\usage{
check_data(x)
}
\arguments{
\item{x}{A Stock Synthesis data list object as read in by \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}}.}
}
\description{
Check that the Stock Synthesis data file looks correct
}
