% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/change_data.r
\name{check_data}
\alias{check_data}
\title{Check that the SS3 data file looks correct}
\usage{
check_data(x)
}
\arguments{
\item{x}{An SS3 data list object as read in by \code{\link[r4ss]{SS_readdat}}.}
}
\description{
Check that the SS3 data file looks correct
}

