% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_q.R
\name{check_q}
\alias{check_q}
\title{Check if desired \eqn{q} parameters exist in control file}
\usage{
check_q(ctl_list, Nfleets, desiredfleets)
}
\arguments{
\item{ctl_list}{A control file read in by \code{\link[r4ss]{SS_readctl}}.}

\item{Nfleets}{The number of fleets in the model.
This can be determined manually from the data file or
using the R object returned from \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}}, of which
\code{Nfleets} is a named element.}

\item{desiredfleets}{A numeric vector specifying which fleets should have
catchability parameters.}
}
\value{
A list with two vectors specifying which fleets to add and which
to remove from the control file.
}
\description{
Check a Stock Synthesis control file to determine if the desired fleets
have q parameters.
}
\examples{
dat <- r4ss::SS_readdat(
  file = dir(
    path = system.file("extdata", "models", "cod-om", package = "ss3sim"),
    pattern = "\\\\.dat",
    full.names = TRUE
  ),
  verbose = FALSE
)
ctl <- r4ss::SS_readctl(
  file = dir(
    path = system.file("extdata", "models", "cod-om", package = "ss3sim"),
    pattern = "\\\\.ctl",
    full.names = TRUE
  ),
  verbose = FALSE,
  use_datlist = TRUE, datlist = dat
)
stopifnot(check_q(ctl, dat[["Nfleets"]], desiredfleets = 1)[["remove"]] == 2)
stopifnot(all(mapply(is.null, check_q(ctl, dat[["Nfleets"]], desiredfleets = 1:2))))
stopifnot(check_q(ctl, dat[["Nfleets"]], desiredfleets = 1:3)[["add"]] == 3)
stopifnot(check_q(ctl, dat[["Nfleets"]], desiredfleets = 2:3)[["remove"]] == 1)

}
\seealso{
\code{\link[=change_q]{change_q()}}
}
