% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expand_scenarios.r
\name{expand_scenarios}
\alias{expand_scenarios}
\title{Create vectors of scenario IDs}
\usage{
expand_scenarios(cases = list(D = 0, E = 0, F = 0, M = 0, R = 0),
  species = c("cod", "fla", "sar"))
}
\arguments{
\item{cases}{A named list of cases. The names in the list are the
case IDs and the values are the case values.}

\item{species}{Vector of 3-letter character IDs designating the
species/stock.}
}
\value{
A character vector of scenario IDs. The case IDs will be
alphabetically sorted.
}
\description{
Create vectors of scenarios from inputs. For passing to
\code{\link{run_ss3sim}}, or \code{\link{get_results_all}}. Default
case values are the base case (\code{0}).
}
\examples{
expand_scenarios()
expand_scenarios(cases = list(D = 0:3, E = 0, F = 0, M = 0, R = 0),
species = "cod")
}
\author{
Cole Monnahan and Sean Anderson
}
\seealso{
\code{\link{run_ss3sim}}, \code{\link{get_results_all}}
}

