% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_across.r
\name{fill_across}
\alias{fill_across}
\title{Fill rows in a matrix}
\usage{
fill_across(mat)
}
\arguments{
\item{mat}{A numeric matrix, typically information extracted from a
data frame but leaving the meta-data behind. Algebraic functions can
then be applied rows or columns of this matrix.}
}
\description{
Fill rows in a matrix
}
