% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fill_across.r
\name{fill_across}
\alias{fill_across}
\title{Fill in matrix across rows of weight-at-age data by interpolation}
\usage{
fill_across(mat, minYear, maxYear)
}
\arguments{
\item{mat}{A matrix}

\item{minYear}{Minimum year}

\item{maxYear}{Maximum year}
}
\description{
Function that fills in matrix across rows of wtatage data by interpolation
Missing Rows are then backfilled
}
\author{
Peter Kuriyama and Allan Hicks
}
\seealso{
\code{\link{sample_lcomp}}, \code{\link{sample_agecomp}},
  \code{\link{fill_across}}
}

