% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_bin.R
\name{get_bin}
\alias{get_bin}
\title{Get SS3 binary/executable location in package}
\usage{
get_bin(bin_name = "ss3_24o_opt")
}
\arguments{
\item{bin_name}{Name of SS3 binary}
}
\value{
The path to an SS3 binary. If using the GitHub version of the
  package, this will be an internal binary. Otherewise, this function
  will search for a version of the binary in your path. See the
  ss3sim vignette.
}
\description{
Get SS3 binary/executable location in package
}
\examples{
get_bin()
}

