% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_bin_info.r
\name{get_bin_info}
\alias{get_bin_info}
\title{Get the parameter values for change_bin}
\usage{
get_bin_info(dat)
}
\arguments{
\item{dat}{A list of sample arguments from all sampling functions}
}
\description{
This function organizes arguments for other functions needed by
\code{change_bin}.
}

