% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case-parsing.r
\name{get_caseval}
\alias{get_caseval}
\title{Take a scenario ID and a case type and return the case number}
\usage{
get_caseval(scenario, case)
}
\arguments{
\item{scenario}{A character object with the cases. E.g.
\code{"M1-F1-D1-R1"}}

\item{case}{The case you want to extract. E.g. \code{"M"}}
}
\description{
Take a scenario ID and a case type and return the case number
}

