% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_fish600_casefolder.r
\name{get_fish600_casefolder}
\alias{get_fish600_casefolder}
\title{Get the folder location of the FISH600 case files}
\usage{
get_fish600_casefolder()
}
\value{
A character object showing the location of the FISH600 case
files in the package \code{extdata} folder.
}
\description{
This function is used by some developers of \pkg{ss3sim} for simulations. This
function links to the "cases" folder in \code{extdata}.
}

