% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_model_folder.r
\name{get_model_folder}
\alias{get_model_folder}
\title{Get the folder location of an included SS3 model configuration}
\usage{
get_model_folder(folder_name)
}
\arguments{
\item{folder_name}{The model folder name. One of \code{"cod-om", "cod-em",
"fla-om", "fla-em", "sar-om", "sar-em"} representing cod, flatfish, and
sardine-like model configurations and operating (om) and estimating model
(em) varieties. See the \pkg{ss3sim} paper or vignette for further details.}
}
\value{
A character object showing the location of the appropriate model
configuration folder in the package \code{extdata} folder.
}
\description{
This function returns the location of one of the built-in model
configurations.
}
\examples{
get_model_folder("cod-em")
}

