% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-results.r
\name{get_nll_components}
\alias{get_nll_components}
\title{Get negative log likelihood (NLL) values from a report file list}
\usage{
get_nll_components(report.file)
}
\arguments{
\item{report.file}{An \code{\link[r4ss]{SS_output}} list for a model
(operating model or estimation model).}
}
\value{
A vector of named numeric values, where \code{"NLL_"} is
appended to the names in the \code{report.file}.
}
\description{
Names of the available NLL components will depend on the version
of the model. Names are native to the estimation framework and all
available components are extracted.
}
\author{
Merrill Rudd
}
