% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-results.r
\name{get_results_all}
\alias{get_results_all}
\title{Extract Stock Synthesis simulation output}
\usage{
get_results_all(
  directory = getwd(),
  overwrite_files = FALSE,
  user_scenarios = NULL,
  type = c("long", "wide"),
  filename_prefix = "ss3sim"
)
}
\arguments{
\item{directory}{The directory which contains scenario folders with
results.}

\item{overwrite_files}{A switch to determine if existing files should be
overwritten, useful for testing purposes or if new iterations are run.}

\item{user_scenarios}{A character vector of scenarios that should be read
in. Default is \code{NULL}, which indicates find all scenario folders in
\code{directory}.}

\item{type}{A character string specifying if you want the results to be
written to the disk and returned as a long or wide data frame, where the
default is \code{"long"}.}

\item{filename_prefix}{A character string specifying a prefix to append to
the filename. Defaults to "ss3sim".}
}
\value{
Returns a list of 3 data frames: \code{scalar}, \code{ts}, and \code{dq}.
Creates two \code{.csv} files in the current working directory,
where the names of those files are based on \code{filename_prefix}
and the default leads to the following:
\code{ss3sim_ts.csv} and \code{ss3sim_scalar.csv}.
}
\description{
This high level function extracts results from Stock Synthesis model runs. Give it a
directory which contains directories for different "scenario" runs, within
which are iterations. It writes two data frames to file:
one for single scalar values (e.g., MSY) and a second
that contains output for each year of the same model (\code{timeseries}, e.g.,
biomass per year). These can always be joined later.
}
\seealso{
Other get-results: 
\code{\link{get_results_derived}()},
\code{\link{get_results_scalar}()},
\code{\link{get_results_scenario}()},
\code{\link{get_results_timeseries}()}
}
\author{
Cole C. Monnahan, Merrill Rudd, Kathryn L. Doering
}
\concept{get-results}
