% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-results.r
\name{get_results_iter}
\alias{get_results_iter}
\title{Get results for 1 iteration}
\usage{
get_results_iter(dir_1_iter = NULL, mod_dirs = NULL, iter_name = NULL)
}
\arguments{
\item{dir_1_iter}{The full or relative path to the SS iteration folder.
Assumed to contain multiple model folders that contain "om" or "em"
(not case sensitive) somewhere in the model file name. If specified,
mod_dirs need not be specified.}

\item{mod_dirs}{The full or relative path to the SS model folders as a
vector of characters. If specified, dir_1_iter need not be specified.}

\item{iter_name}{Name of the iteration, which will be appended to the
dataframes . Defaults to NULL, in which case the iter_name will be the
folder name of dir_1_iter or the folder name 1 level up from the first
mod_dirs specified}
}
\value{
A list of 3 data frames called scalar, timeseries, and
derived (for derived quantities). These lists contain information for
multiple model runs (estimation models and operating models) for 1
iteration.
}
\description{
Get results for 1 iteration
}
\author{
Kathryn Doering
}
