% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-results.r
\name{get_results_scalar}
\alias{get_results_scalar}
\title{Extract scalar quantities from a model run.}
\usage{
get_results_scalar(report.file)
}
\arguments{
\item{report.file}{An \code{\link[r4ss]{SS_output}} list for a model
(operating model or estimation model).}
}
\description{
Extract scalar quantities from an \code{\link[r4ss]{SS_output}} list from a model run.
Returns a data.frame of the results (a single row) which can be rbinded later.
}
\author{
Cole Monnahan; Merrill Rudd
}
\seealso{
Other get-results: \code{\link{get_results_all}};
  \code{\link{get_results_derived}};
  \code{\link{get_results_scenario}};
  \code{\link{get_results_timeseries}}
}

