% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-results.r
\name{get_results_scenario}
\alias{get_results_scenario}
\title{Extract Stock Synthesis simulation results for one scenario}
\usage{
get_results_scenario(scenario, directory = getwd(), overwrite_files = FALSE)
}
\arguments{
\item{scenario}{A single character giving the scenario from which to
extract results.}

\item{directory}{The directory which contains the scenario folder.}

\item{overwrite_files}{A boolean (default is \code{FALSE}) for whether to delete
any files previously created with this function. This is intended to be
used if iterations were added since the last time it was called, or any
changes were made to this function.}
}
\description{
Extract results from all iterations inside a supplied scenario folder.
The function writes the following \code{.csv} files to the scenario folder
\enumerate{
\item scalar metrics with one value per iteration
(e.g. \eqn{R_0}, \eqn{h}),
\item a \code{timeseries} data frame ('ts') which contains multiple values per iteration
(e.g.,  \eqn{SSB_y} for a range of years \eqn{y}), and
\item residuals on the log scale from the surveys across all iterations;
this functionality is currently disabled and not tested.
}
}
\seealso{
\code{\link[=get_results_all]{get_results_all()}} loops through these \code{.csv} files and
combines them together into a single "final" data frame.

Other get-results: 
\code{\link{get_results_all}()},
\code{\link{get_results_derived}()},
\code{\link{get_results_scalar}()},
\code{\link{get_results_timeseries}()}
}
\author{
Cole Monnahan and Kathryn Doering
}
\concept{get-results}
