% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-results.r
\name{get_results_timeseries}
\alias{get_results_timeseries}
\title{Extract time series from a model run.}
\usage{
get_results_timeseries(report.file)
}
\arguments{
  \item{report.file}{An \code{SS_output} list for a model
  (operating model or estimation model).}
}
\description{
Extract time series from an \code{SS_output} list from a
model run. Returns a data.frame of the results for SSB,
recruitment and effort by year.
}
\author{
Cole Monnahan
}
\seealso{
Other get-results: \code{\link{get_results_all}};
\code{\link{get_results_scalar}};
\code{\link{get_results_scenario}}
}

