% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sigma.r
\name{get_sigmar}
\alias{get_sigmar}
\title{Get Variability About Recruitment Deviations (\eqn{\sigma_R})}
\usage{
get_sigmar(om)
}
\arguments{
\item{om}{The name of the operating model, which should be the prefix of
the \code{.ctl} file, e.g., "myOM".
A full directory can be specified with the the prefix of the file name but
leaving off the '.ctl' portion.}
}
\description{
Use the name of the operating model to open the ctl file and obtain the
INIT value for sigmaR (recruitment deviations sigma)
}
\author{
Kelli F. Johnson
}
