% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss_version.R
\name{get_ss_ver_dl}
\alias{get_ss_ver_dl}
\title{Get the Stock Synthesis version (either 3.24 or 3.30) from a list object}
\usage{
get_ss_ver_dl(dat_list)
}
\arguments{
\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}
}
\description{
Get the Stock Synthesis version from a data list, \code{dat_list},
that was a originally read in using \code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}}.
}
