% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ss_version.R
\name{get_ss_ver_file}
\alias{get_ss_ver_file}
\title{Get ths ss version (either 3.24 or 3.30) from an ss file (but not forecast.ss)}
\usage{
get_ss_ver_file(file)
}
\arguments{
\item{file}{Input SS3 control file, either a starter, control, or data SS file.}
}
\description{
# Get the SS version from the top line in an SS file.
as done in \code{\link[r4ss]{SS_readdat}}.
}
