% Generated by roxygen2 (4.0.2): do not edit by hand
\name{id_scenarios}
\alias{id_scenarios}
\title{Identify ss3sim scenarios within a directory}
\usage{
id_scenarios(directory)
}
\arguments{
  \item{directory}{The directory which contains scenario
  folders with results.}
}
\description{
Identify ss3sim scenarios within a directory
}
\author{
Merrill Rudd
}

