% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-results.r
\name{id_scenarios}
\alias{id_scenarios}
\title{Identify ss3sim scenarios within a directory}
\usage{
id_scenarios(directory)
}
\arguments{
\item{directory}{The directory which contains scenario folders with
results.}
}
\value{
A character vector of folders
}
\description{
Identify ss3sim scenarios within a directory
}
\author{
Merrill Rudd
}

