% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot_re.R
\name{plot_re}
\alias{plot_re}
\title{Plot relative error}
\usage{
plot_re(data, re, x, row_dat, column_dat, mar = c(2.5, 0.5, 0.5, 0.15),
  oma = c(3.5, 4.5, 3.5, 1), cex = 0.6, ...)
}
\description{
Plot relative error
}
\examples{
library("dplyr")
d <- scalar_dat
d <- calculate_re(d)
d <- d \%>\% select(D, E, SSB_MSY_re, depletion_re, SSB_Unfished_re) \%>\%
  filter(!D \%in\% "D100", !E \%in\% c("E100", "E101"))
d <- d \%>\% reshape2::melt(value.name = "re")
plot_re(data = d, re = "re", x = "variable", row_dat = "D", column_dat = "E")
}

