% Generated by roxygen2 (4.0.1): do not edit by hand
\name{profile_fmsy}
\alias{profile_fmsy}
\title{Determine Fmsy for a given operating model}
\usage{
profile_fmsy(om_in, results_out, simlength = 100, start = 0, end = 1.5,
  by_val = 0.01)

profile_fmsy(om_in, results_out, simlength = 100, start = 0, end = 1.5,
  by_val = 0.01)
}
\arguments{
  \item{om_in}{A directory for an \code{ss3sim} operating
  model.}

  \item{results_out}{A directory to place the results}

  \item{simlength}{Number of total years the \code{om_in}
  is set to run.}

  \item{start}{Lower fishing mortality level}

  \item{end}{Upper fishing mortality level}

  \item{by_val}{Interval in which you wish to increment the
  fishing mortality level}

  \item{om_in}{A directory for an \code{ss3sim} operating
  model.}

  \item{results_out}{A directory to place the results}

  \item{simlength}{Number of total years the \code{om_in}
  is set to run.}

  \item{start}{Lower fishing mortality level}

  \item{end}{Upper fishing mortality level}

  \item{by_val}{Interval in which you wish to increment the
  fishing mortality level}
}
\description{
Runs an operating model over a range of fishing mortality
levels to to determine the profile of F values from which
Fmsy can be determined.

Runs an operating model over a range of fishing mortality
levels to to determine the profile of F values from which
Fmsy can be determined.
}
\examples{
d <- system.file("extdata", package = "ss3sim")
omfolder <- paste0(d, "/models/cod-om")

fmsy.val <- fmsy_profile(om_in = omfolder,
                         results_out = "c:/ss/fmsy",
                         start = 0.1, end = 0.2, by_val = 0.001)
d <- system.file("extdata", package = "ss3sim")
omfolder <- paste0(d, "/models/cod-om")

fmsy.val <- fmsy_profile(om_in = omfolder,
                         results_out = "c:/ss/fmsy",
                         start = 0.1, end = 0.2, by_val = 0.001)
}

