% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_ss3model.r
\name{rename_ss3_files}
\alias{rename_ss3_files}
\title{Rename SS3-version-specific files}
\usage{
rename_ss3_files(path, ss_bin, extensions)
}
\arguments{
\item{path}{The path to the folder with the files.}

\item{ss_bin}{A character value giving the SS binary name}

\item{extensions}{A character vector of file extensions to rename without
periods preceding the values.}
}
\description{
Rename SS3-version-specific files
}
\author{
Sean C. Anderson
}

