% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ss3model.r
\name{run_ss3model}
\alias{run_ss3model}
\title{Run an operating or estimation model in a directory}
\usage{
run_ss3model(
  dir,
  admb_options = "",
  hess = FALSE,
  ignore.stdout = TRUE,
  admb_pause = 0.05,
  show.output.on.console = FALSE,
  ...
)
}
\arguments{
\item{dir}{A character string specifying the path to the folder with the
results from a stock assessment model run. Paths are passed without a
terminal slash.}

\item{admb_options}{Any additional options to pass to the SS command.}

\item{hess}{Calculate the Hessian on estimation model runs?}

\item{ignore.stdout}{Passed to \code{system}. If \code{TRUE} then ADMB
output is not printed on screen. This will be slightly faster. Set to
\code{FALSE} to help with debugging.}

\item{admb_pause}{A length of time (in seconds) to pause after running the
simulation model. This can be necessary on certain computers where file
writing can be slightly delayed. For example, on computers where the files
are written over a network connection. If the output files haven't
finished writing before \R starts looking for the output then the
simulation will crash with an error about missing files. The default
value is set to \code{0.01} seconds, just to be safe.}

\item{show.output.on.console}{Logical: passed on to
\code{system}.}

\item{...}{Anything else to pass to \code{system}.}
}
\description{
Call Stock Synthesis (SS) to generate data or fit model to data.
The appropriate version of SS is called after determining if the
user is on Unix or Windows. This lower-level
function is meant to be called by higher level functions such as
\code{\link{run_ss3sim}}, \code{\link{ss3sim_base}}, or your own custom
function.
}
\details{
ss3sim requires you to place the SS executable in your
path. See the vignette \code{vignette("introduction", package = "ss3sim")} for details on
this process. The executables themselves can be downloaded from github
\url{https://github.com/ss3sim/ss3sim/tree/master/inst/bin} or the
SS website
\url{https://vlab.noaa.gov/web/stock-synthesis/document-library/-/document_library/0LmuycloZeIt/view/5042555}.
If the github version of ss3sim is installed, rather than the CRAN version,
then the executables are automatically downloaded and called from their
stored location rather than from the path variable.
}
\seealso{
\code{\link{ss3sim_base}}, \code{\link{run_ss3sim}}
}
\author{
Sean C. Anderson
}
