% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_calcomp.R
\name{sample_calcomp}
\alias{sample_calcomp}
\title{Sample conditional age-at-length (CAL) data}
\usage{
sample_calcomp(dat_list, outfile = NULL, fleets = c(1, 2), years, Nsamp)
}
\arguments{
\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}} in the \pkg{r4ss} package. Make sure
you select option \code{section=2}.}

\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the 
file takes time and computing resources.}

\item{fleets}{*A vector of integers specifying which fleets to include.
The order of the fleets pertains to the input order of other arguments.
An entry of \code{fleets=NULL} leads to zero samples for any fleet.}

\item{years}{*A list the same length as \code{fleets} giving the years as
numeric vectors. If no fleet collected samples, keep the value to
\code{years=NULL}.}

\item{Nsamp}{*A numeric list of the same length as \code{fleets}.
Either single values or vectors of the same length as the number of
years can be passed through. Single values are repeated for all
years. If no fleet collected samples, keep the value to
\code{Nsamp=NULL}.}
}
\value{
A modified \code{.dat} file if \code{!is.null(outfile)}. A list object
containing the modified \code{.dat} file is returned invisibly.
}
\description{
Samples conditional age-at-length (CAL) data from expected values of length
proportions and expected values of age proportions (conditional on length)
from the operating model (OM) and writes the samples to file for use by the
estimation model (EM).
}
\details{
This function takes a \code{data.SS_new} file from an OM containing
 expected values of length proportions and age proportions, conditional on
 length. First, sample from true length proportions, using the length comp sample
 sizes, to get realistic sample sizes for age bins given their lengths. Next,
 use these sample sizes and the expected values of age proportions
 (conditional on length) to sample for realistic age proportions. If no fish
 are sampled, for a row of age proportions, then that row is discarded. A
 value of NULL for fleets indicates to delete the data, so if the EM is used
 with \code{\link{run_ss3sim}}, the case file should be named \code{calcomp}.
 Only the multinomial distribution is currently implemented, so this function
 cannot be used with the dirichlet distribution. Note also that this sampling
 procedure represents the case where 1) lengths are sampled randomly, 2) fish
 are lengthed and placed into bins, and 3) a subset of lengthed fish are
 aged, where a constant proportion from each length bin are selected for
 aging. This does NOT represent response stratified sampling where a subset
 of lengthed fish are aged, and a constant number from each length bin is
 selected for aging, although these data could also be put into a Stock
 Synthesis model as "Conditional Age at Length."
}
\note{
This function is only reliable when using multinomial length
 compositions for the fleet(s) with conditional age at length sampling. The
 real-valued length compositions resulting from the Dirichlet distribution
 cause difficulties in the sampling code. See the vignette for more
 information.
}
\section{Which arguments to specifiy in case files}{

All function argument descriptions that start with an asterisk (*) will be passed
through the case files to \code{\link{run_ss3sim}}. If one of these arguments
is not specified in a case file, then a value of \code{NULL} will be passed,
which may or may not be an appropriate value. Other arguments will be ignored
if specified.
}

\seealso{
Other sampling functions: 
\code{\link{clean_data}()},
\code{\link{sample_agecomp}()},
\code{\link{sample_index}()},
\code{\link{sample_lcomp}()},
\code{\link{sample_mlacomp}()},
\code{\link{sample_wtatage}()}
}
\author{
Cole Monnahan, Kotaro Ono
}
\concept{sampling functions}
