% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_calcomp.R
\name{sample_calcomp}
\alias{sample_calcomp}
\title{Sample conditional age-at-length (CAL) data and write to file for use by
the EM.}
\usage{
sample_calcomp(datfile, outfile, fleets = c(1, 2), years, write_file = TRUE,
  Nsamp)
}
\arguments{
\item{datfile}{An SS data object as read in from
\code{\link[r4ss]{SS_readdat}} in the \pkg{r4ss} package. Make sure
you select option \code{section=2}.}

\item{outfile}{A character string of the new file name to be created.
Must end in \code{.dat} or equal \code{wtatage.ss}.}

\item{fleets}{*A numeric vector giving the fleets to be used. This order also
pertains to other arguments. A missing value excludes that fleet from
\code{outfile} (i.e. it turns it off so no samples are written). If none of
the fleet collected samples, keep the value to \code{fleets=NULL}.}

\item{years}{*A list the same length as \code{fleets} giving the years as
numeric vectors. If no fleet collected samples, keep the value to
\code{years=NULL}.}

\item{write_file}{A switch for whether to write \code{outfile} to
disk. Can be turned off to speed up testing or exploration of the
function. The new data are returned invisibly, as in the examples
below.}

\item{Nsamp}{*A numeric list of the same length as \code{fleets}.
Either single values or vectors of the same length as the number of
years can be passed through. Single values are repeated for all
years. If no fleet collected samples, keep the value to
\code{Nsamp=NULL}.}
}
\value{
A modified \code{.dat} file if \code{write_file=TRUE}. A list object
containing the modified \code{.dat} file is returned invisibly.
}
\description{
Sample conditional age-at-length (CAL) data and write to file for use by
the EM.
}
\details{
Take a \code{data.SS_new} file containing expected values and
sample from true lengths, using length comp sample sizes, to get
realistic sample sizes for age bins given a length. If no fish are
sampled then that row is discarded. A value of NULL for fleets indicates
to delete the data so the EM
If used with \code{\link{run_ss3sim}} the case file should be named
\code{calcomp}.
}
\section{Which arguments to specifiy in case files}{

All function argument descriptions that start with an asterisk (*) will be passed
through the case files to \code{\link{run_ss3sim}}. If one of these arguments
is not specified in a case file, then a value of \code{NULL} will be passed,
which may or may not be an appropriate value. Other arguments will be ignored
if specified.
}
\author{
Cole Monnahan, Kotaro Ono
}
\seealso{
Other sampling functions: \code{\link{clean_data}};
  \code{\link{sample_agecomp}}; \code{\link{sample_index}};
  \code{\link{sample_lcomp}}; \code{\link{sample_mlacomp}};
  \code{\link{sample_wtatage}}
}

