% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_index.r
\name{sample_index}
\alias{sample_index}
\title{Sample the biomass with observation error}
\usage{
sample_index(
  dat_list,
  outfile = NULL,
  fleets,
  years,
  sds_obs,
  sds_out,
  seas = list(1)
)
}
\arguments{
\item{dat_list}{A Stock Synthesis  data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the
file takes time and computing resources.}

\item{fleets}{*A vector of integers specifying which fleets to include.
The order of the fleets pertains to the input order of other arguments.
An entry of \code{fleets=NULL} leads to zero samples for any fleet.}

\item{years}{*A list the same length as \code{fleets} giving the years as
numeric vectors. If no fleet collected samples, keep the value to
\code{years=NULL}.}

\item{sds_obs, sds_out}{A list the same length as \code{fleets} specifying the
standard deviation of the observation error used for the sampling and
the value used in the returned output. Thus, \code{sds_obs} is what is actually
used to sample and \code{sds_out} can be used to test what happens when the
input values to your model are biased. If \code{sds_out} is missing, then
\code{sds_obs} will be used for the output as well as the input.
List elements should be
either single numeric values or numeric vectors the same length as the
number of years sampled for each given fleet.
Single values are repeated for all years.
See details for more information, particularly for the equations.}

\item{seas}{A list the same length as \code{fleets}. The list should contain
either single numeric values or numeric vectors, where the length of each
vector matches the length of each vector present in \code{years}.
Single values are repeated for all years.
This parameter operates the same as \code{sds_obs} but it specifies the
season rather than the standard deviation of the observations.}
}
\value{
A modified \code{.dat} file if \code{!is.null(outfile)}. A list object
containing the modified \code{.dat} file is returned invisibly.
}
\description{
Create an index of abundance sampled from the expected
available biomass for specified fleets in specified years.
}
\details{
Samples are generated using the following equation:
\deqn{
  B_y*exp(stats::rnorm(1, 0, sds_obs)-sds_obs^2/2)
},
where \eqn{B_y} is the expected biomass in year y and
\eqn{sds_obs} is the standard deviation of the normal distribution or
the standard error of the \eqn{log_e(B_y)}.
For the error term, this is the same
parameterization that is used in Stock Synthesis.
More details can be found in the
\href{https://nmfs-stock-synthesis.github.io/doc/SS330_User_Manual_release.html#indices}{section on indices in the Stock Synthesis manual}
The second term in the equation adjusts the random samples so their expected
value is \eqn{B_y}, i.e., the log-normal bias correction.

If you only know the coefficient of variation (\eqn{CV}), then
the input error can be approximated using \eqn{\sqrt{log_e(1+CV^{2})}}.
Where, \eqn{CV} is assumed to be constant with mean changes in biomass.
The lognormal distribution can be approximated by a proportional
distribution or normal distribution only when the variance is low, i.e.,
\eqn{CV < 0.50} or log standard deviation of 0.22.
}
\examples{
# Find the example data location:
set.seed(3)
dat_list <- r4ss::SS_readdat(
  file = file.path(
    system.file("extdata", "example-om", package = "ss3sim"),
    "ss3_expected_values.dat"
  ),
  verbose = FALSE
)
# Look at expected values for the index data
# fleet 2, every other year from 76 to 100
# dat_list$CPUE
sam_yrs <- seq(76, 100, by = 2)
ex1 <- sample_index(dat_list,
  outfile = NULL,
  fleets = 2,
  seas = list(unique(
    dat_list[["CPUE"]][dat_list[["CPUE"]][, "index"] == 2, "seas"]
  )),
  years = list(sam_yrs),
  sds_obs = list(seq(0.001, 0.1, length.out = length(sam_yrs)))
)
\dontshow{
testthat::expect_equivalent(
  ex1[["CPUE"]][1:2, "obs"],
  c(1472202421, 1554321845)
)
}
\dontrun{
ex1$CPUE
# could sample from less years, but not more:
ex2 <- sample_index(dat_list,
  outfile = NULL,
  fleets = 2,
  seas = list(unique(
    dat_list[["CPUE"]][dat_list[["CPUE"]][, "index"] == 2, "seas"]
  )),
  years = list(sam_yrs[c(-1, -2)]),
  sds_obs = list(seq(0.001, 0.1, length.out = length(sam_yrs) - 2))
)
ex2$CPUE
# sd can be fixed across years:
ex3 <- sample_index(dat_list,
  outfile = NULL,
  fleets = 2,
  seas = list(unique(
    dat_list[["CPUE"]][dat_list[["CPUE"]][, "index"] == 2, "seas"]
  )),
  years = list(sam_yrs),
  sds_obs = list(0.01)
)
ex3$CPUE
# If fleet 1 also had expected values in the index that you wanted to sample:
testthat::expect_error(
  ex4 <- sample_index(dat_list,
    outfile = NULL,
    fleets = c(1, 2),
    years = list(sam_yrs, sam_yrs),
    sds_obs = list(0.01, 0.01)
  )
)
# sd in the returned file can be different than what is used to sample:
ex5 <- sample_index(dat_list,
  outfile = NULL,
  fleets = 2,
  seas = list(unique(
    dat_list[["CPUE"]][dat_list[["CPUE"]][, "index"] == 2, "seas"]
  )),
  years = list(sam_yrs),
  sds_obs = list(0.01),
  sds_out = list(0.20)
)
ex5$CPUE
testthat::expect_true(all(ex5[["CPUE"]][["se_log"]] == 0.2))
}
}
\seealso{
Other sampling functions: 
\code{\link{clean_data}()},
\code{\link{sample_agecomp}()},
\code{\link{sample_calcomp}()},
\code{\link{sample_catch}()},
\code{\link{sample_discard}()},
\code{\link{sample_lcomp}()},
\code{\link{sample_mlacomp}()},
\code{\link{sample_wtatage}()}
}
\author{
Cole Monnahan, Kotaro Ono
}
\concept{sampling functions}
