% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mlacomp.R
\name{sample_mlacomp}
\alias{sample_mlacomp}
\title{\strong{BETA VERSION} Sample mean length (size-)-at-age data and write to file for use by the EM}
\usage{
sample_mlacomp(
  dat_list,
  outfile,
  ctl_file_in,
  fleets = 1,
  Nsamp,
  years,
  mean_outfile = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the
file takes time and computing resources.}

\item{ctl_file_in}{A path to the control file, output from an OM, containing
the OM parameters for growth. These values are used to determine the
uncertainty about size for fish sampled in each age bin.}

\item{fleets}{*A vector of integers specifying which fleets to include.
The order of the fleets pertains to the input order of other arguments.
An entry of \code{fleets=NULL} leads to zero samples for any fleet.}

\item{Nsamp}{*A numeric list of the same length as \code{fleets}.
Either single values or vectors of the same length as the number of
years can be passed through. Single values are repeated for all
years. If no fleet collected samples, keep the value to
\code{Nsamp=NULL}.}

\item{years}{*A list the same length as \code{fleets} giving the years as
numeric vectors. If no fleet collected samples, keep the value to
\code{years=NULL}.}

\item{mean_outfile}{A path to write length and age data for external
estimation of parametric growth. If \code{NULL} no file will be written.
This file is used by \code{\link[=change_e]{change_e()}} to externally estimate growth
parameters. Filename must contain "vbgf" to be used by \code{\link[=change_e]{change_e()}}.
Also, if "remove" is included in the filename, the mean length at age data
will be removed from the \code{.dat} file and not be available to the EM.}

\item{verbose}{Logical value whether or not diagnostic information from
\pkg{r4ss} functions should be printed to the screen. Default is FALSE.}
}
\value{
A modified \code{.dat} file if \code{!is.null(outfile)}. A list object
containing the modified \code{.dat} file is returned invisibly.
}
\description{
\strong{BETA VERSION} Sample mean length (size-)-at-age data and write to file for use by the EM
}
\details{
\strong{This function is in beta and untested. Use with caution.}
Take a \code{data.SS_new} file, read in by \pkg{r4ss} function
\code{\link[r4ss:SS_readdat]{r4ss::SS_readdat()}} containing observed values, and
sample from the observed ages to get realistic proportions for the number
of fish in each age bin, then use the mean size-at-age and CV for growth to
generate random samples of size, which are then averaged to get mean
length-at-age values. These values are then written to file for the
EM.
}
\seealso{
Other sampling functions: 
\code{\link{clean_data}()},
\code{\link{sample_agecomp}()},
\code{\link{sample_calcomp}()},
\code{\link{sample_catch}()},
\code{\link{sample_discard}()},
\code{\link{sample_index}()},
\code{\link{sample_lcomp}()},
\code{\link{sample_wtatage}()}
}
\author{
Cole Monnahan, Kelli Johnson
}
\concept{sampling functions}
