% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sample_mlacomp.R
\name{sample_mlacomp}
\alias{sample_mlacomp}
\title{Sample mean length (size-)-at-age data and write to file for use by the EM.}
\usage{
sample_mlacomp(datfile, outfile, ctlfile, fleets = 1, Nsamp, years,
  write_file = TRUE, mean_outfile = NULL, verbose = TRUE)
}
\arguments{
\item{datfile}{An SS data object as read in from
\code{\link[r4ss]{SS_readdat}} in the \pkg{r4ss} package. Make sure
you select option \code{section=2}.}

\item{outfile}{A character string of the new file name to be created.
Must end in \code{.dat} or equal \code{wtatage.ss}.}

\item{ctlfile}{A path to the control file, outputed from an OM, containing
the OM parameters for growth. These values are used to determine the
uncertainty about size for fish sampled in each age bin.}

\item{fleets}{*A numeric vector giving the fleets to be used. This order also
pertains to other arguments. A missing value excludes that fleet from
\code{outfile} (i.e. it turns it off so no samples are written). If none of
the fleet collected samples, keep the value to \code{fleets=NULL}.}

\item{Nsamp}{*A numeric list of the same length as \code{fleets}.
Either single values or vectors of the same length as the number of
years can be passed through. Single values are repeated for all
years. If no fleet collected samples, keep the value to
\code{Nsamp=NULL}.}

\item{years}{*A list the same length as \code{fleets} giving the years as
numeric vectors. If no fleet collected samples, keep the value to
\code{years=NULL}.}

\item{write_file}{A switch for whether to write \code{outfile} to
disk. Can be turned off to speed up testing or exploration of the
function. The new data are returned invisibly, as in the examples
below.}

\item{mean_outfile}{A path to write length and age data for external
estimation of parametric growth. If \code{NULL} no file will be written.
This file is used by \code{change_e} to externally estimate growth
parameters. Filename must contain "vbgf" to be used by \code{change_e}.
Also, if "remove" is included in the filename, the mean length at age data
will be removed from the \code{.dat} file and not be available to the EM.}

\item{verbose}{Logical value whether or not diagnostic information from
\pkg{r4ss} functions should be printed to the screen. Default is FALSE.}
}
\value{
A modified \code{.dat} file if \code{write_file=TRUE}. A list object
containing the modified \code{.dat} file is returned invisibly.
}
\description{
Sample mean length (size-)-at-age data and write to file for use by the EM.
}
\details{
Take a \code{data.SS_new} file, read in by \pkg{r4ss} function
  \code{\link[r4ss]{SS_readdat}} containing observed values, and
  sample from the observed ages to get realistic proportions for the number
  of fish in each age bin, then use the mean size-at-age and CV for growth to
  generate random samples of size, which are then averaged to get mean
  length-at-age values. These values are then written to file for the
  EM.
}
\section{Which arguments to specifiy in case files}{

All function argument descriptions that start with an asterisk (*) will be passed
through the case files to \code{\link{run_ss3sim}}. If one of these arguments
is not specified in a case file, then a value of \code{NULL} will be passed,
which may or may not be an appropriate value. Other arguments will be ignored
if specified.
}
\examples{
temp_path <- file.path(tempdir(), "ss3sim-test")
dir.create(temp_path, showWarnings = FALSE)
wd <- getwd()
setwd(temp_path)
d <- system.file("extdata/models/cod-om", package = "ss3sim")
dat_in <- file.path(d, "codOM.dat")
datfile <- r4ss::SS_readdat(dat_in, verbose = FALSE)
datfile <- change_fltname(datfile)
datfile <- change_data(datfile, outfile = NULL, write_file = FALSE,
  fleets = 1, years = 1990:2010, types = c("age", "mla"))
datfile <- change_fltname(datfile)
ctlfile <- file.path(d, "codOM.ctl")

out <- sample_mlacomp(datfile, outfile = NULL, ctlfile = ctlfile,
                      fleets = 1, Nsamp = 30, years = list(1992),
                      verbose = FALSE, mean_outfile = "test.csv", write_file = FALSE)

setwd("..")
unlink("ss3sim-test", recursive = TRUE)
setwd(wd)
}
\author{
Cole Monnahan, Kelli Johnson
}
\seealso{
Other sampling functions: \code{\link{clean_data}};
  \code{\link{sample_agecomp}};
  \code{\link{sample_calcomp}}; \code{\link{sample_index}};
  \code{\link{sample_lcomp}}; \code{\link{sample_wtatage}}
}

