% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_wtatage.R
\name{sample_wtatage}
\alias{sample_wtatage}
\title{Sample empirical weight-at-age data and write to file for use by the EM}
\usage{
sample_wtatage(
  wta_file_in,
  outfile,
  dat_list,
  ctl_file_in,
  years,
  fill_fnc = fill_across,
  fleets,
  cv_wtatage = NULL
)
}
\arguments{
\item{wta_file_in}{The file to read weight-at-age from. Specifically to get the
age-0 weight-at-age. This is typically \code{wtatage.ss_new}.}

\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the 
file takes time and computing resources.}

\item{dat_list}{An SS data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated give all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{ctl_file_in}{A path to the control file, output from an OM, containing
the OM parameters for growth and weight/length relationship. These values
are used to determine the uncertainty about weight for fish sampled in each
age bin. Commonly \code{control.ss_new}}

\item{years}{*A list the same length as \code{fleets} giving the years as
numeric vectors. If no fleet collected samples, keep the value to
\code{years=NULL}.}

\item{fill_fnc}{A function to fill in missing values (ages and years). The
resulting weight-at-age file will have values for all years and ages.One
function is \code{fill_across}.}

\item{fleets}{*A vector of integers specifying which fleets to include.
The order of the fleets pertains to the input order of other arguments.
An entry of \code{fleets=NULL} leads to zero samples for any fleet.}

\item{cv_wtatage}{A user specified CV for growth. Default is \code{NULL}.}
}
\value{
A modified \code{.wtatage.ss} file if \code{!is.null(outfile)}. A list
  object containing the modified \code{.wtatage.ss} file is returned invisibly.
}
\description{
Take a \code{data.SS_new} file containing expected values and sample from
true ages to get realistic proportions for the number of fish in each age
bin, then use the mean size-at-age and CV for growth to generate random
samples of size, which are then converted to weight and averaged to get mean
weight-at-age values. Missing ages and years are filled according to a
specified function. These matrices are then written to file for the EM. By
calling this function, \pkg{ss3sim} will turn on the empirical weight-at-age
function (set maturity option to 5) automatically. See
\code{\link{ss3sim_base}} for more details on how that is implemented.
}
\seealso{
\code{\link{fill_across}}

Other sampling functions: 
\code{\link{clean_data}()},
\code{\link{sample_agecomp}()},
\code{\link{sample_calcomp}()},
\code{\link{sample_index}()},
\code{\link{sample_lcomp}()},
\code{\link{sample_mlacomp}()}
}
\author{
Cole Monnahan, Allan Hicks, Peter Kuriyama
}
\concept{sampling functions}
