% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_wtatage.R
\name{sample_wtatage}
\alias{sample_wtatage}
\title{Sample empirical weight-at-age data}
\usage{
sample_wtatage(
  wta_file_in,
  outfile,
  dat_list,
  ctl_file_in,
  years,
  fill_fnc = fill_across,
  fleets,
  cv_wtatage = NULL
)
}
\arguments{
\item{wta_file_in}{The file to read weight-at-age from. Specifically to get the
age-0 weight-at-age. This is typically \code{wtatage.ss_new}.}

\item{outfile}{A character string specifying the file name to use
when writing the information to the disk. The string must include
the proper file extension. No file is written using the default value
of \code{NULL}, which leads to increased speed because writing the
file takes time and computing resources.}

\item{dat_list}{A Stock Synthesis  data list object as read in from
\code{\link[r4ss]{SS_readdat}}.
Be sure to correctly specify which section of the data file you want
to work with when reading it in using the \code{section} argument.
Where, \code{section = 1} reads in the input values used to run the model
and \code{section = 2} reads in the expected values generated given all the
input to the OM. \code{section = 3} is not used within ss3sim, but this
section provides bootstrapped data sets that have been sampled internally
within SS.}

\item{ctl_file_in}{A path to the control file, output from an OM, containing
the OM parameters for growth and weight/length relationship. These values
are used to determine the uncertainty about weight for fish sampled in each
age bin. Commonly \code{control.ss_new}}

\item{years}{*A list the same length as \code{fleets} giving the years as
numeric vectors. If no fleet collected samples, keep the value to
\code{years=NULL}.}

\item{fill_fnc}{A function to fill in missing values (ages and years). The
resulting weight-at-age file will have values for all years and ages.One
function is \code{\link[=fill_across]{fill_across()}}.}

\item{fleets}{*A vector of integers specifying which fleets to include.
The order of the fleets pertains to the input order of other arguments.
An entry of \code{fleets=NULL} leads to zero samples for any fleet.}

\item{cv_wtatage}{A user specified coefficient of variation (CV) for growth.
Default is \code{NULL}.}
}
\value{
A modified \code{.wtatage.ss} file if \code{!is.null(outfile)}. A list
object containing the modified \code{.wtatage.ss} file is returned invisibly.
}
\description{
In Stock Synthesis, empirical weight-at-age data can be used to
read empirical body weight for the population from each fleet.
This data removes the use of growth parameters from the EM because
weights are assigned to each age internally rather than from the growth parameters,
from which spawning biomass/fecundity can be determined.
These values are not data in the sense they have a likelihood but are generated from samples.
Sampling empirical weight-at-age data from the expected values takes many steps.
}
\details{
The steps for sampling empirical weight-at-age are as follows:
\itemize{
\item Sample from the expected ages to get realistic proportions for
the number of fish in each age bin.
\item Use the mean size-at-age and coefficient of variation for growth to
generate random samples of size,
which are then converted to weight and averaged to get mean weight-at-age values.
\item Fill in missing ages and years.
\item Write the information to the appropriate files.
\item Turn on weight-at-age data in Stock Synthesis by setting the maturity option to 5.
}
}
\seealso{
\itemize{
\item \code{\link[=fill_across]{fill_across()}}
\item \code{\link[=ss3sim_base]{ss3sim_base()}}
}

Other sampling functions: 
\code{\link{clean_data}()},
\code{\link{sample_agecomp}()},
\code{\link{sample_calcomp}()},
\code{\link{sample_catch}()},
\code{\link{sample_discard}()},
\code{\link{sample_index}()},
\code{\link{sample_lcomp}()},
\code{\link{sample_mlacomp}()}
}
\author{
Cole Monnahan, Allan Hicks, Peter Kuriyama
}
\concept{sampling functions}
