% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sample_wtatage.R
\name{sample_wtatage}
\alias{sample_wtatage}
\title{Sample empirial weight-at-age data and write to file for use by the EM}
\usage{
sample_wtatage(infile, outfile, datfile, ctlfile, years, fill_fnc,
  write_file = TRUE)
}
\arguments{
  \item{infile}{The file to read weight-at-age from.
  Specifically to get the age-0 weight-at-age. This is
  typically \code{wtatage.ss_new}.}

  \item{outfile}{The file to write the created
  weight-at-age matrices to be read in by the estimation
  model. Commonly \code{wtatage.ss}.}

  \item{datfile}{A path to the data file, outputed from an
  OM, containing the true age distributions (population
  bins). This file is read in and then used to determine
  how many fish of each age bin are to be sampled. Commonly
  \code{data.ss_new}}

  \item{ctlfile}{A path to the control file, outputed from
  an OM, containing the OM parameters for growth and
  weight/length relationship. These values are used to
  determine the uncertainty about weight for fish sampled
  in each age bin. Commonly \code{control.ss_new}}

  \item{years}{*A list of vectors for each fleet indicating
  the years that are sampled for weight-at-age. There must
  be a corresponding age composition for that year in
  \code{datfile}.}

  \item{fill_fnc}{*A function to fill in missing values
  (ages and years). The resulting weight-at-age file will
  have values for all years and ages.One function is
  \code{fill_across}.}

  \item{write_file}{Logical to determine if \code{outfile}
  will be written.}
}
\value{
A modified \code{.dat} file if \code{write_file=TRUE}. A
list object containing the modified \code{.dat} file is
returned invisibly.
}
\description{
Take a \code{data.SS_new} file containing expected values
and sample from true ages to get realistic proportions for
the number of fish in each age bin, then use the mean
size-at-age and CV for growth to generate random samples of
size, which are then converted to weight and averaged to
get mean weight-at-age values. Missing ages and years are
filled according to a specified function. These matrices
are then written to file for the EM. By calling this
function, \pkg{ss3sim} will turn on the empirical
weight-at-age function (set maturity option to 5)
automatically. See \code{\link{ss3sim_base}} for more
details on how that is implemented.
}
\section{Which arguments to specifiy in case files}{
  All function argument descriptions that start with an
  asterisk (*) will be passed through the case files to
  \code{\link{run_ss3sim}}. If one of these arguments is
  not specified in a case file, then a value of \code{NULL}
  will be passed, which may or may not be an appropriate
  value. Other arguments will be ignored if specified.
}
\author{
Cole Monnahan, Allan Hicks, Peter Kuriyama
}
\seealso{
\code{\link{sample_lcomp}, \link{sample_agecomp}},
\code{\link{fill_across}}
}

