% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_ss3model.r
\name{sanitize_admb_options}
\alias{sanitize_admb_options}
\title{Check admb options to make sure there aren't flags there shouldn't
be}
\usage{
sanitize_admb_options(x, exclude = "-nohess")
}
\arguments{
\item{x}{The admb options}

\item{exclude}{A character object (not a vector)}
}
\description{
Check admb options to make sure there aren't flags there shouldn't
be
}
\author{
Sean C. Anderson
}

