% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_scenarios.R
\name{setup_scenarios_defaults}
\alias{setup_scenarios_defaults}
\title{Set up a generic scenario}
\usage{
setup_scenarios_defaults()
}
\value{
A data frame with the minimal information needed to run a scenario.
}
\description{
Create a data frame of scenario inputs for a generic simulation that will
run within ss3sim. Users can add more arguments, but the scenario will run
without changing the returned value.
}
\author{
Kelli Faye Johnson
}
