% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/case-parsing.r
\name{substr_r}
\alias{substr_r}
\title{Substring from right}
\usage{
substr_r(x, n)
}
\arguments{
\item{x}{A character object}

\item{n}{The number of characters from the right to extract}
}
\description{
Substring from right
}
\references{
http://stackoverflow.com/questions/7963898/extracting-the-last-n-characters-from-a-string-in-r
}

