% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/growth_estimation.R
\name{vbgf_func}
\alias{vbgf_func}
\title{Predict length given VBGF parameters}
\usage{
vbgf_func(L1, L.inf, k, ages, a3)
}
\arguments{
\item{L1}{mean length at youngest age which is well sampled in the data (a3)}

\item{L.inf}{Length at infinity}

\item{k}{von bertalanffy growth rate parameter}

\item{ages}{vector of ages in the data for which you want to predict mean
length-at-age}

\item{a3}{youngest age which is well sampled in the data}
}
\value{
a vector of lengths predicted which correspond to the input ages
  vector.
}
\description{
External estimation procedure for von Bertalanffy growth.
}

