% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_comps.R
\name{weight_comps}
\alias{weight_comps}
\title{Weight the composition data in the EM}
\usage{
weight_comps(
  method = c("MI", "Francis", "DM"),
  init_run = FALSE,
  niters_weighting = 1,
  dir,
  bias_adjust = FALSE,
  hess_always = FALSE,
  fleets,
  ...
)
}
\arguments{
\item{method}{Options for the weighting method are
"MI" (for MacAllister-Ianelli), "Francis", or "DM" (for Dirichlet
multinomial). For now, only 1 option can be used.}

\item{init_run}{A logical value specifying if the model should be run before
iterative re-weighting. Only relevant to \code{method = c("MI", "Francis")}
because they require initial estimates that are compared to a theoretical
distribution external to SS.}

\item{niters_weighting}{The number of times you want to tune the model.}

\item{dir}{A character string specifying the path to the folder with the
results from a stock assessment model run. Paths are passed without a
terminal slash.}

\item{bias_adjust}{Run bias adjustment first?}

\item{hess_always}{If \code{TRUE}, the Hessian will always be calculated.
If \code{FALSE}, the Hessian will only be calculated for
bias-adjustment runs thereby saving time.}

\item{fleets}{A vector of numeric values specifying which fleets you want to
tune.}

\item{...}{Anything else to pass to \code{\link{run_ss3model}}.}
}
\description{
Weight age- or length-composition data in the estimation model (EM) using a
tuning method (MacAllister-Ianelli or Francis) or a distribution
(Dirichlet multinomial). This is essentially a wrapper around
r4ss::SS_tune_comps
}
